% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliaryFunctions.R
\name{addDesign}
\alias{addDesign}
\title{Add design variables to hyperframe}
\usage{
addDesign(hypFrame, desMat, designVec)
}
\arguments{
\item{hypFrame}{The hyperframe}

\item{desMat}{The design matrix}

\item{designVec}{The design vector}
}
\value{
The hyperframe with design variables added
}
\description{
Add design variables to hyperframe
}
