% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliaryFunctions.R
\name{addTabObs}
\alias{addTabObs}
\title{Add tables with gene counts to the hyperframe, presort by gene and x-ccordinate
and add design varibales}
\usage{
addTabObs(hypFrame, desMat)
}
\arguments{
\item{hypFrame}{The hyperframe}

\item{desMat}{The design matrix}
}
\value{
The hyperframe with tabObs added
}
\description{
Add tables with gene counts to the hyperframe, presort by gene and x-ccordinate
and add design varibales
}
