% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalWeightFun.R
\name{addWeights}
\alias{addWeights}
\title{Add weights to PI matrix}
\usage{
addWeights(piMat, pi, obj)
}
\arguments{
\item{piMat}{The exisitng matrix with PI and design variables}

\item{pi}{character string indicating the desired pim as outcome
in the linear mixed model}

\item{obj}{The result of a call to buildWeightFunction()}
}
\value{
The matrix with the weights column added
}
\description{
Add weights to PI matrix
}
