% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildDataFrame.R
\name{buildDataFrame}
\alias{buildDataFrame}
\title{Build a data frame for a single gene and measure to prepare
for mixed model building}
\usage{
buildDataFrame(
  obj,
  gene,
  pi = c("nn", "nnPair", "edge", "midpoint", "nnCell", "nnPairCell")
)
}
\arguments{
\item{obj}{The result of a call to buildWeightFunction()}

\item{gene}{A character string indicating the desired gene or gene pair}

\item{pi}{character string indicating the desired pim as outcome
in the linear mixed model}
}
\value{
A dataframe
}
\description{
Build a data frame for a single gene and measure to prepare
for mixed model building
}
\examples{
example(addWeightFunction, "spatrans")
dfUniNN <- buildDataFrame(yangObj, gene = 'SmVND2', pi = 'nn')
# Example analysis with linear mixed model
library(lmerTest)
# Use sum coding for day to maintain interpretability of the intercept
mixedMod <- lmer(pi - 0.5 ~ day + (1 | root),
    weight = weight, data = dfUniNN,
    contrasts = list('day' = 'contr.sum')
)
summary(mixedMod)
# Evidence for aggregation
}
