% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buidDfMM.R
\name{buildDfMM}
\alias{buildDfMM}
\title{Build a data frame for a single gene and measure to prepare for mixed model building}
\usage{
buildDfMM(pimRes, gene, pi)
}
\arguments{
\item{pi}{}
}
\description{
Build a data frame for a single gene and measure to prepare for mixed model building
}
