% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildFormula.R
\name{buildFormula}
\alias{buildFormula}
\title{Build a formula from different components}
\usage{
buildFormula(Formula, fixedVars, randomVars, outcome = "pi - 0.5")
}
\arguments{
\item{Formula}{A formula. If not supplied or equals NULL, will be overridden}

\item{fixedVars, randomVars}{Character vectors with fixed and random variables}

\item{outcome}{A character vector describing the outcome}
}
\value{
A formula
}
\description{
Build a formula from different components
}
\details{
Random intercepts are assumed for the random effects, if more
complicated designs are used, do supply your own formula.
}
\seealso{
\link{fitLMMs},\link[stats]{formula}
}
