% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildMoransIDataFrame.R
\name{buildMoransIDataFrame}
\alias{buildMoransIDataFrame}
\title{Build a data frame with Moran's I as outcome variable}
\usage{
buildMoransIDataFrame(pi, piMat, weightMats)
}
\arguments{
\item{pi}{character string indicating the desired PI as outcome}

\item{piMat}{A data frame. Will be constructed if not provided}
}
\value{
A modified data frame
}
\description{
Build a data frame with Moran's I as outcome variable
}
