% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildMoransIWeightMat.R
\name{buildMoransIWeightMat}
\alias{buildMoransIWeightMat}
\title{Build a weight matrix for Moran's I calculations}
\usage{
buildMoransIWeightMat(coordList, numNNs)
}
\arguments{
\item{coordList}{A list of coordinate matrices}
}
\value{
A sparse matrix of weights for Moran's I statistic, of equal value
for the numNNs nearest neighbours and zero otherwise
}
\description{
Build a weight matrix for Moran's I calculations
}
