% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildWeightFunction.R
\name{buildWeightFunction}
\alias{buildWeightFunction}
\title{Build a weight function based on the data}
\usage{
buildWeightFunction(
  pimRes,
  pi = c("nn", "allDist", "nnPair", "allDistPair"),
  hypFrame,
  designVars,
  maxObs = 1e+06,
  maxFeatures = 1000,
  ...
)
}
\arguments{
\item{pimRes}{The object containing all}

\item{pi}{A character string, the PI required}

\item{hypFrame}{The hyperframe with all the data}

\item{designVars}{A character vector containing all design factors (both fixed and random), that are also present as variables in hypFrame}

\item{maxObs, maxFeatures}{The maximum number of observations respectively features for fitting the weight function. See details}

\item{...}{Additional arguments passed on to the scam::scam function}
}
\value{
A fitted scam object, which can be used to
}
\description{
Build a weight function based on the data
}
\details{
For computational and memory reasons, for large datasets the trend fitting
is restricted to a subset of the data through the maxObs and maxFeatures parameters.

The scam functions fits a decreasing spline of the variance as a function of the number of observations
}
\examples{
data(Yang)
hypYang = buildHyperFrame(Yang, coordVars = c("x", "y"),
designVar = c("day", "root", "section"))
yangPims = estPims(hypYang, pis = c("nn", "nnPair"))
#First Build the weight function
wf <- buildWeightFunction(yangPims, pi = "nn", hypFrame = hypYang, designVars = c("day", "root"))
}
\seealso{
[buildDfMM]
}
