% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcAllDistPI.R
\name{calcAllDistPI}
\alias{calcAllDistPI}
\title{Estimate the probabilistic index for the collection of all distances}
\usage{
calcAllDistPI(pSub, ecdfAll)
}
\arguments{
\item{pSub}{The subset point pattern wiht only one feature}

\item{ecdfAll}{The empirical cumulative distribution function of all distances under the null}
}
\value{
The probabilistic index for all distances
}
\description{
Estimate the probabilistic index for the collection of all distances
}
