% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcBiPIs.R
\name{calcBiPIs}
\alias{calcBiPIs}
\title{Calculate bivaraiet PIs}
\usage{
calcBiPIs(
  p,
  pis,
  null,
  nSims,
  ecdfs,
  nSub,
  ecdfAll,
  features,
  manyPairs,
  verbose,
  allowManyGenePairs,
  piPair
)
}
\arguments{
\item{p}{The point pattern}

\item{pis}{the Probabilitstic indices to be estimated}

\item{null}{A character vector, indicating how the null distribution is defined. See details.}

\item{nSims}{The number of Monte-Carlo simulations used to determine the null distribution if null = "CSR}

\item{ecdfs}{Ecdfs per point}

\item{nSub}{The number of events in the subsampled pattern}

\item{ecdfAll}{Overall ecdf}

\item{features}{A character vector, for which features should the probablilistic indices be calculated?}
}
\value{
A matrix of bivariate PIs
}
\description{
Calculate bivaraiet PIs
}
