% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcEdgeDist.R
\name{calcEdgeDistPI}
\alias{calcEdgeDistPI}
\title{Estimate the PI for the distance to a fixed region or point}
\usage{
calcEdgeDistPI(pSub, edge, ecdfAll)
}
\arguments{
\item{pSub}{The subset point pattern containing only a single gene}

\item{edge}{The psp object to which the distances are calculated}

\item{point}{The fixed point to which the distance is to be calculated}
}
\value{
The estimated probabilistic index
}
\description{
Estimate the PI for the distance to a fixed region or point
}
