% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPointDistPI.R
\name{calcFixedPointDistPI}
\alias{calcFixedPointDistPI}
\title{Estimate the PI for the distance to a fixed region of interest}
\usage{
calcFixedPointDistPI(pSub, point, ecdfAll)
}
\arguments{
\item{point}{The fixed point to which the distance is to be calculated}
}
\description{
Estimate the PI for the distance to a fixed region of interest
}
