% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIndividualPIs.R
\name{calcIndividualPIs}
\alias{calcIndividualPIs}
\title{Calculate individual PI entries}
\usage{
calcIndividualPIs(
  p,
  tabObs,
  pis,
  pSubLeft,
  owins,
  centroids,
  null,
  features,
  ecdfAll,
  ecdfsCell
)
}
\arguments{
\item{p}{The point pattern}

\item{tabObs}{A table of observed gene frequencies}

\item{pis}{The probabilistic indices to be estimated}

\item{pSubLeft}{The subsampled overall point pattern}

\item{owins, centroids}{The list of windows corresponding to cells,
and their centroids}

\item{null}{A character vector, indicating how the null distribution is
defined. See details.}

\item{features}{A character vector, for which features should the
probabilistic indices be calculated?}
}
\value{
A list over the features with all PI entries
}
\description{
Calculate individual PI entries
}
\details{
For the single-feature nearest neighbour distances, the average
is already calculated
}
