% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcNNpi.R
\name{calcNNPI}
\alias{calcNNPI}
\title{Estimate the PI for the nearest neighbour distances}
\usage{
calcNNPI(pSub, p, null, ecdfs, n, ecdfAll)
}
\arguments{
\item{pSub}{The subset point pattern with only one feature}

\item{p}{The point pattern}

\item{null}{A character vector, indicating how the null distribution is defined. See details.}

\item{ecdfs}{A list of empirical distribution functions, one for each event}

\item{n}{the number of subsampled distances}

\item{ecdfAll}{The empirical cumulative distribution function of all distances under the null}
}
\value{
The estimated probabilistic index
}
\description{
Estimate the PI for the nearest neighbour distances
}
