% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPointDistPI.R
\name{calcPointDistPI}
\alias{calcPointDistPI}
\title{Estimate the PI for the distance to a fixed region of interest}
\usage{
calcPointDistPI(pSub, point, ecdfAll)
}
\arguments{
\item{pSub}{The subset point pattern containing only a single gene}

\item{point}{The fixed point to which the distance is to be calculated}

\item{ecdfAll}{the cumulative distribution function under the null}
}
\description{
Estimate the PI for the distance to a fixed region of interest
}
