% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcUniPIs.R
\name{calcUniPIs}
\alias{calcUniPIs}
\title{Calculate univariate PIs for all features}
\usage{
calcUniPIs(
  p,
  pis,
  verbose,
  ecdfsCell,
  owins,
  tabObs,
  null,
  cd,
  nPointsAll,
  nSub,
  ecdfAll,
  features,
  centroids
)
}
\arguments{
\item{p}{The point pattern}

\item{pis}{The probabilistic indices to be estimated}

\item{verbose}{A boolean, should verbose output be printed}

\item{ecdfsCell}{ecdfs within the cells}

\item{owins, centroids}{The list of windows corresponding to cells,
and their centroids}

\item{tabObs}{A table of observed gene frequencies}

\item{null}{A character vector, indicating how the null distribution is
defined. See details.}

\item{cd}{A matrix of cross-distances}

\item{nPointsAll}{How many points to subsample or simulate to calculate
overall interpoint distance and distance to point.
This parameter is a strong driver of memory and cpu usage.}

\item{nSub}{The number of events in the subsampled pattern}

\item{ecdfAll}{Overall ecdf}

\item{features}{A character vector, for which features should the
probabilistic indices be calculated?}
}
\value{
PIs for every feature
}
\description{
Calculate univariate PIs for all features
}
