% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkAttr.R
\name{checkAttr}
\alias{checkAttr}
\title{Check if the required pi's are present in the object}
\usage{
checkAttr(pimRes, pi)
}
\arguments{
\item{pimRes}{The result of the PI calculation}

\item{pi}{A character string indicating the desired PI}
}
\value{
Throws an error when the PIs are not found, otherwise returns invisible
}
\description{
Check if the required pi's are present in the object
}
