% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkAttr.R
\name{checkAttrPimRes}
\alias{checkAttrPimRes}
\title{Check if the required pi's are present in the pim result object}
\usage{
checkAttrPimRes(pimRes, pi)
}
\arguments{
\item{pimRes}{The result of the PI calculation}

\item{pi}{A character string indicating the desired PI}
}
\value{
Throws an error when the PIs are not found, otherwise returns invisible
}
\description{
Check if the required pi's are present in the pim result object
}
