% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkAttr.R
\name{checkAttrWf}
\alias{checkAttrWf}
\title{Check if the required pi's are present in the weight function}
\usage{
checkAttrWf(wf, pi)
}
\arguments{
\item{wf}{the weight function object}
}
\description{
Check if the required pi's are present in the weight function
}
