% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkPi.R
\name{checkFeatures}
\alias{checkFeatures}
\title{Check if features are present in hyperframe}
\usage{
checkFeatures(hypFrame, features)
}
\arguments{
\item{hypFrame}{A hyperframe}

\item{features}{A character vector, for which features should the
probabilistic indices be calculated?}
}
\value{
Throws error when features not found
}
\description{
Check if features are present in hyperframe
}
