% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkPi.R
\name{checkPi}
\alias{checkPi}
\title{Check if the required PI's are present in the object}
\usage{
checkPi(x, pi)
}
\arguments{
\item{x}{The result of the PI calculation, or a weighting function}

\item{pi}{A character string indicating the desired PI}
}
\value{
Throws an error when the PIs are not found, otherwise returns invisible
}
\description{
Check if the required PI's are present in the object
}
