% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertToOwins.R
\name{convertToOwins}
\alias{convertToOwins}
\title{Converta list of differently formatted windows to owins}
\usage{
convertToOwins(windows, coords, ...)
}
\arguments{
\item{windows}{The list of windows}

\item{coords}{The names of the coordinates, if the windows are given as sets of coordinates}

\item{...}{Further arguments passed onto read functions}
}
\value{
A list of owins
}
\description{
Converta list of differently formatted windows to owins
}
