% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{crossdistWrapper}
\alias{crossdistWrapper}
\title{A wrapper for C-functions calculating cross-distance matrix fast}
\usage{
crossdistWrapper(x, y)
}
\arguments{
\item{x, y}{the matrices or point patterns between which to calculate the
cross distances}
}
\value{
a matrix of cross distances
}
\description{
A wrapper for C-functions calculating cross-distance matrix fast
}
