% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliaryFunctions.R
\name{dropDiagonal}
\alias{dropDiagonal}
\title{Remove the diagonal from a square matrix}
\usage{
dropDiagonal(x)
}
\arguments{
\item{x}{The matrix from which to remove the diagonal}
}
\value{
A matrix with the same number of rows as x and one column less
}
\description{
Remove the diagonal from a square matrix
}
\details{
The diagonal elements are removed, and the elements right from the
diagonal are shifted left to form a new matrix with one column less
}
