% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecdfFast.R
\name{ecdfFast}
\alias{ecdfFast}
\title{A fast version that makes an ecdf based on an ordered vector and the cumulative probability function}
\usage{
ecdfFast(x, cumDens)
}
\arguments{
\item{x}{The ordered vector}

\item{cumDens}{Corresponding cumulative density}
}
\value{
The ecdf function
}
\description{
A fast version that makes an ecdf based on an ordered vector and the cumulative probability function
}
