% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecdfPreSort.R
\name{ecdfPreSort}
\alias{ecdfPreSort}
\title{A fast version that makes an ecdf based on an presorted vector}
\usage{
ecdfPreSort(x)
}
\arguments{
\item{x}{The ordered vector}
}
\value{
The ecdf function
}
\description{
A fast version that makes an ecdf based on an presorted vector
}
