% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Eng}
\alias{Eng}
\alias{EngRois}
\title{Spatial transcriptomics data of mouse fibroblast cells}
\format{
\enumerate{
\item \strong{Eng} A data frame with variables
\describe{
  \item{x,y}{Molecule coordinates}
  \item{gene}{Character vector with gene identities}
 \item{experiment,fov}{Design variables}
}
\item \strong{EngRois} A list of list of regions of interest (ROIs)
}
}
\source{
\doi{10.1038/s41586-019-1049-y}
}
\usage{
data(Eng)
}
\description{
Single-molecule spatial transcriptomics seqFISH+ data containing measurements of 10,000 genes in NIH/3T3 mouse fibroblast cells by
\insertCite{Eng2019}{spatrans}. Molecule locations, gene identity and design
variables are included, a subset of eight most expressed genes is included in the package,
and the dataset was subsampled to 100,000 observations for memory reasons.
In addition, a list of regions of interest (rois) is given describing the cell boundaries.
}
\references{
\insertAllCited{}
}
\keyword{datasets}
