% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estGradients.R
\name{estGradientsSingle}
\alias{estGradientsSingle}
\title{A wrapper function for the estimation of the different gradients on a hyperframe for a single feature}
\usage{
estGradientsSingle(
  hypFrame,
  gradients,
  fixedForm,
  randomForm,
  fixedFormSimple,
  effects = NULL,
  ...
)
}
\arguments{
\item{hypFrame}{A hyperframe}

\item{gradients}{The gradients indices to be estimated}

\item{fixedForm, randomForm, fixedFormSimple}{Formulae for fixed effects, random effects and fixed effects without
slopes respectively}

\item{effects}{Character vector of fixed and random effects}

\item{...}{Passed onto fitGradient}
}
\value{
A list contraining
\item{overall}{Overall gradients}
\item{cell}{Gradients within the cell}
}
\description{
Workhorse for \link{estGradients} on a single feature
}
\seealso{
\link{estGradients}
}
