% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estPis.R
\name{estPisSingle}
\alias{estPisSingle}
\title{A wrapper function for the different probabilistic indices (PIs), applied to individual point patterns}
\usage{
estPisSingle(
  p,
  pis,
  null,
  tabObs,
  owins = NULL,
  centroids = NULL,
  window = p$window,
  loopFun = "bplapply",
  features,
  nPointsAll,
  nPointsAllWithinCell,
  nPointsAllWin,
  minDiff
)
}
\arguments{
\item{p}{The point pattern}

\item{pis}{The probabilistic indices to be estimated}

\item{null}{A character vector, indicating how the null distribution is
defined. See details.}

\item{tabObs}{A table of observed gene frequencies}

\item{owins, centroids}{The list of windows corresponding to cells,
and their centroids}

\item{window}{An window of class owin, in which events can occur}

\item{loopFun}{The function to use to loop over the features.
Defaults to bplapply except when looping over features within cells}

\item{features}{A character vector, for which features should the
probabilistic indices be calculated?}

\item{nPointsAll, nPointsAllWithinCell}{How many points to subsample or simulate to calculate the
overall distance distribution under the null hypothesis.
The second argument (nPointsAllWithinCell) applies to within cell calculations, where a lower number usually suffises.}

\item{nPointsAllWin}{How many points to subsample or simulate
to calculate distance to cell edge or centroid distribution}

\item{minDiff}{An integer, the minimum number of events from other genes
needed for calculation of background distribution of distances}
}
\value{
A list of data frames with estimated PIs per gene and/or gene pair:
\item{pointDists}{PIs for pointwise distances overall}
\item{windowDists}{PIs for distances to cell wall or centroid}
\item{withinCellDists}{PIs for pointwise distances within cell}
}
\description{
A wrapper function for the different probabilistic indices (PIs), applied to individual point patterns
}
\seealso{
\links{estPis}
}
