% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findOverlap.R
\name{findOverlap}
\alias{findOverlap}
\title{Try to find overlap between list of windows, and throw error when found}
\usage{
findOverlap(owins)
}
\arguments{
\item{owins}{the list of windows}
}
\value{
Throws an error when overlap found, otherwise returns invisible
}
\description{
Try to find overlap between list of windows, and throw error when found
}
