% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitGradient.R
\name{fitGradient}
\alias{fitGradient}
\title{Fit a gradient to a point pattern}
\usage{
fitGradient(
  hypFrame,
  fixedForm,
  randomForm,
  fixedFormSimple,
  returnModel = FALSE,
  silent,
  ...
)
}
\arguments{
\item{hypFrame}{the hyperframe}

\item{returnModel}{A boolean, should the entire model be returned?
Otherwise the p-value is returned}

\item{silent}{A boolean, should error messages from ppm.ppp be printed?}

\item{...}{passed onto spatstat.model::ppm}
}
\value{
A list with p-value and coefficients,
or a mppm model when returnModel is true
}
\description{
A Poisson process is fitted to the data assuming exponential relationship
between x and y variables and intensity
}
\seealso{
\link{estGradients}
}
