% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitLMMs.R
\name{fitLMMsAll}
\alias{fitLMMsAll}
\title{A wrapper function for linear mixed model fitting for all PIs}
\usage{
fitLMMsAll(obj, pis = obj$pis, ...)
}
\arguments{
\item{obj}{The result object}

\item{pis}{Optional, the pis required. Defaults to all pis in the object}

\item{...}{Passed onto fitLMMs}
}
\value{
A list of fitted objects
}
\description{
A wrapper function for linear mixed model fitting for all PIs
}
