% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitPiModel.R
\name{fitPiModel}
\alias{fitPiModel}
\title{Fit a linear model for an individual gene and PI combination}
\usage{
fitPiModel(Formula, dff, contrasts, Control, MM, Weight = NULL)
}
\arguments{
\item{Formula}{A formula; if not supplied it will be constructed
from the fixed and random variables}

\item{dff}{The dataframe}

\item{contrasts}{The contrasts to be used, see \link{model.matrix}}

\item{Control}{Control parameters}

\item{MM}{A boolean, should a mixed model be tried}

\item{Weight}{A weight variable}
}
\value{
A fitted model
}
\description{
Fit a linear model for an individual gene and PI combination
}
\seealso{
\link{fitLMMsSingle}
}
