% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitPiModel.R
\name{fitSingleLmmModel}
\alias{fitSingleLmmModel}
\title{Take an existing frame, add outcome and weight ad fit lmer model}
\usage{
fitSingleLmmModel(ff, y, wts = NULL)
}
\arguments{
\item{ff}{}

\item{y}{outcome vector}

\item{wts}{weights vector}
}
\value{
A fitted lmer model
}
\description{
Take an existing frame, add outcome and weight ad fit lmer model
}
