% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcWindowDistPI.R
\name{getApproxRanks}
\alias{getApproxRanks}
\title{Approximate the ranks given and ecdf}
\usage{
getApproxRanks(ecdf, obs)
}
\arguments{
\item{ecdf}{An empirical density function}

\item{obs}{A vector of observations}
}
\value{
The approximate ranks
}
\description{
Approximate the ranks given and ecdf
}
