% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliaryFunctions.R
\name{getCoordsMat}
\alias{getCoordsMat}
\title{Extract coordinates from a point pattern or data frame}
\usage{
getCoordsMat(x)
}
\arguments{
\item{x}{the point pattern, dataframe or matrix}
}
\value{
the matrix of coordinates
}
\description{
Extract coordinates from a point pattern or data frame
}
