% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux.R
\name{getElDist}
\alias{getElDist}
\title{Extract elements from a distance matrix by index}
\usage{
getElDist(distMat, i, j)
}
\arguments{
\item{distMat}{the distance matrix of class 'dist'}

\item{i, j}{Indices for which entries are extracted}
}
\description{
Extract elements from a distance matrix by index
}
\details{
See documentation of the stats::dist function
}
