% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{getElement}
\alias{getElement}
\title{Extract en element from a matrix or vector}
\usage{
getElement(x, e)
}
\arguments{
\item{x}{the matrix or vector}

\item{e}{The column or element name}
}
\value{
The desired element
}
\description{
Extract en element from a matrix or vector
}
