% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFeatures.R
\name{getFeatures}
\alias{getFeatures}
\title{Extract all unique features from an object}
\usage{
getFeatures(x)
}
\arguments{
\item{x}{A hyperframe or a results list containing a hyperframe}
}
\value{
A vector of features.
}
\description{
Extract all unique features from an object
}
\details{
If x is a hyperFrame, all of its features are returned. If it is a result from estPis,
only features with estimated PIs are returned.
}
\examples{
data(Yang)
hypYang <- buildHyperFrame(Yang,
    coordVars = c("x", "y"),
    imageVars = c("day", "root", "section")
)
head(getFeatures(hypYang))
}
