% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGp.R
\name{getGp}
\alias{getGp}
\title{Helper function to get gene pair or order reversed}
\usage{
getGp(x, gp, Collapse = "--")
}
\arguments{
\item{x}{The object in which to look}

\item{gp}{A character string describing the gene pair}

\item{Collapse}{The character separating the gene pair}
}
\value{
The element sought
}
\description{
Helper function to get gene pair or order reversed
}
