% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliaryFunctions.R
\name{getHypFrame}
\alias{getHypFrame}
\title{Extract the hyperframe}
\usage{
getHypFrame(x)
}
\arguments{
\item{x}{The hyperframe, or list containing one}
}
\value{
the hyperframe
}
\description{
Extract the hyperframe
}
