% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliaryFunctions.R
\name{getN}
\alias{getN}
\title{Return the number of observations used to fit the ecdf}
\usage{
getN(ecdf)
}
\arguments{
\item{ecdf}{The ecdf function}
}
\value{
The number of observations
}
\description{
Return the number of observations used to fit the ecdf
}
