% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildDataFrame.R
\name{getPiAndWeights}
\alias{getPiAndWeights}
\title{Build a matrix with pi and weights}
\usage{
getPiAndWeights(obj, gene, pi, piMat, prepMat, prepTab)
}
\arguments{
\item{obj}{A results object. For distances to fixed objects, the result of a call to \link{estPis};
for nearest neighbour distances, the result of a call to \link{addWeightFunction}}

\item{gene}{A character string indicating the desired gene or gene pair (genes separated by double hyphens)}

\item{pi}{character string indicating the desired PI}

\item{piMat}{A data frame. Will be constructed if not provided, for internal use.}

\item{prepMat, prepTab}{Preconstructed objects to avoid looping over genes. For internal use mainly}
}
\value{
A matrix of two columns: pi estimate and weights
}
\description{
Build a matrix with pi and weights
}
