% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDesignVars.R
\name{getPPPvars}
\alias{getPPPvars}
\title{Extract variables from point patterns}
\usage{
getPPPvars(
  x,
  exclude = c("tabObs", "centroids", "owins", "ppp", "pimRes", "inSeveralCells")
)
}
\arguments{
\item{x}{The results list}

\item{exclude}{variables to exclude}
}
\value{
A vector of variables
}
\description{
Extract variables from point patterns
}
