% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estGradients.R
\name{getPvaluesGradient}
\alias{getPvaluesGradient}
\title{Extract the p-values for gradient fitting}
\usage{
getPvaluesGradient(res, gradient, method = "BH")
}
\arguments{
\item{res}{The fitted gradients}

\item{gradient}{The gradient to be extracted, a character vector equal to
"overall" or "cell".}

\item{method}{Method of multiplicity correction, see \link{p.adjust}.
Defaults to Benjamini-Hochberg}
}
\value{
A vector of p-values
}
\description{
Extract the p-values for gradient fitting
}
\examples{
example(estGradients, "smoppix")
pVals = getPvaluesGradient(engGrads, "cell")
}
