% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sund.R
\name{makeDesignVar}
\alias{makeDesignVar}
\title{Make design variable by combining different design variables}
\usage{
makeDesignVar(x, designVars, sep = "_")
}
\arguments{
\item{sep}{}
}
\description{
Make design variable by combining different design variables
}
