% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moransI.R
\name{moransI}
\alias{moransI}
\title{Calculate the Moran's I test statistic for spatial autocorrelation}
\usage{
moransI(x, W)
}
\arguments{
\item{x}{A vector of outcomes}

\item{W}{The matrix of weights, with dimensions equal to the lenght of x}
}
\value{
The normalized Moran's I statistic
}
\description{
The normalized Moran's I test statistic is calculated, with standard normal
distribution under the null hypothesis
}
\note{
The implementation is inspired on the one from ape::Moran.I, but more
 bare-bones faster and using less memory.
 @details Calculations are only correct for weight matrices as prepared by
 buildMoransIWeightMat!#'
}
