% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotExplore.R
\name{plotExplore}
\alias{plotExplore}
\title{Plot the hyperframe for exploratory purposes}
\usage{
plotExplore(hypFrame, features = attr(hypFrame, "features")[seq_len(8)], ppps)
}
\arguments{
\item{hypFrame}{The hyperframa}

\item{features}{A small number of features to be fitted. Defaults to the first 5}

\item{ppps}{The rownames or indices of the point patterns to be plotted. Defaults to the first 9.}
}
\description{
Plot the hyperframe for exploratory purposes
}
\examples{
data(Yang)
hypYang = buildHyperFrame(Yang, coordVars = c("x", "y"),
designVar = c("day", "root", "section"))
plotExplore(hypYang)
}
