% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTopResult.R
\name{plotTopResult}
\alias{plotTopResult}
\title{Plot the most significant findings for a certain PI}
\usage{
plotTopResult(
  hypFrame,
  results,
  pi,
  effect = "Intercept",
  smallPI = TRUE,
  sigLevel = 0.05,
  numFeats = 2,
  piThreshold = 0.5,
  ...
)
}
\arguments{
\item{hypFrame}{The hyperframe}

\item{results}{The results frame}

\item{pi}{The probabilistic index}

\item{effect}{The name of the effect}

\item{smallPI}{A boolean, should features with PI smaller than piThreshold be shown? See details}

\item{sigLevel}{The significance level}

\item{numFeats}{The number of features to plot}

\item{piThreshold}{The threshold for PI, a minimum effect size}

\item{...}{passed onto plotting functions plotCells or plotExplore}
}
\value{
A plot from plotCells or plotExplore, throws a warning when no features meet the criteria
}
\description{
Plot the most significant findings for a certain PI
}
\details{
If smallPI is set to TRUE, features with aggregation, colocalization and vicinity to cell boundary or centroid are shown.
If smallPI is set to FALSE, features with regularity, antilocalization and remoteness from cell boundary or centroid are shown.
}
\examples{
example(fitLMMs, "spatrans")
plotTopResult(hypYang, lmmModels, "nn")
}
