% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotWf.R
\name{plotWf}
\alias{plotWf}
\title{Title}
\usage{
plotWf(wf, se = FALSE, ...)
}
\arguments{
\item{wf}{The fitted scam object}

\item{se}{A boolean, should standard errors be plotted?}

\item{...}{Passed onto the plot.scam function for 1D splines}
}
\description{
Title
}
