% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rowSumsWrapper.R
\name{rowSumsWrapper}
\alias{rowSumsWrapper}
\title{A wrapper for C-functions calculating rowSums(x <y) with y a vector and x a
(possible big) matrix}
\usage{
rowSumsWrapper(x, y)
}
\arguments{
\item{x, y}{the matrices or point patterns between which to calculate the
cross distances}
}
\value{
a matrix of cross distances
}
\description{
A wrapper for C-functions calculating rowSums(x <y) with y a vector and x a
(possible big) matrix
}
