% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{selfName}
\alias{selfName}
\title{Name a character vector after itself}
\usage{
selfName(x)
}
\arguments{
\item{x}{The vector to be names}
}
\value{
the named vector
}
\description{
Name a character vector after itself
}
\examples{
selfName(LETTERS[1:5])
}
