% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoppix-package.R
\docType{package}
\name{smoppix}
\alias{smoppix-package}
\alias{smoppix}
\title{smoppix: Analyze Single Molecule Spatial Omics Data Using the Probabilistic Index}
\description{
Test for univariate and bivariate spatial patterns in spatial omics data with single-molecule resolution. The tests implemented allow for analysis of nested designs and are automatically calibrated to different biological specimens. Tests for aggregation, colocalization, gradients and vicinity to cell edge or centroid are provided.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/sthawinke/smoppix}
  \item Report bugs at \url{https://github.com/sthawinke/smoppix/issues}
}

}
\author{
\strong{Maintainer}: Stijn Hawinkel \email{stijn.hawinkel@psb.ugent.be} (\href{https://orcid.org/0000-0002-4501-5180}{ORCID})

}
\keyword{internal}
