% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{sortGp}
\alias{sortGp}
\title{Sort feature pairs alphabetically}
\usage{
sortGp(features)
}
\arguments{
\item{features}{}
}
\value{
A character vector of the same length as the features, with pairs sorted
}
\description{
Sort feature pairs alphabetically
}
