% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{splitWindow}
\alias{splitWindow}
\title{Split a number of plots into rows and colums}
\usage{
splitWindow(x)
}
\arguments{
\item{x}{The number of plots}
}
\value{
A vector of length 2 with required number of rows and colums
}
\description{
Split a number of plots into rows and colums
}
