% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliaryFunctions.R
\name{subSampleP}
\alias{subSampleP}
\title{Subsample a point pattern when it is too large}
\usage{
subSampleP(p, nSims, returnId = FALSE)
}
\arguments{
\item{p}{The point pattern}

\item{nSims}{The maximum size}

\item{returnId}{A boolean, should the id of the sampled elements be returned?}
}
\value{
A point pattern, subsampled if necessary
}
\description{
Subsample a point pattern when it is too large
}
