% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliaryFunctions.R
\name{which.max.early}
\alias{which.max.early}
\title{A faster way to find the first element in a vector larger than a fixed value,
 by stopping early. This function is only faster than which.max is the match
 indeed comes early, as for nearest neighbour in a set of all distances.}
\usage{
which.max.early(x, a)
}
\arguments{
\item{x}{An ordered vector}

\item{a}{A value to be queried}
}
\value{
The index
}
\description{
A faster way to find the first element in a vector larger than a fixed value,
 by stopping early. This function is only faster than which.max is the match
 indeed comes early, as for nearest neighbour in a set of all distances.
}
