% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeToXlsx.R
\name{writeToXlsx}
\alias{writeToXlsx}
\title{Write linear modelling results to excel worksheet}
\usage{
writeToXlsx(obj, file, overwrite = FALSE)
}
\arguments{
\item{obj}{The results of linear model fitting}

\item{file}{The file to write the results to}
}
\value{
Returns invisible with a message when writing operation successful,
otherwise throws an error.
}
\description{
The results are written to different tabs for every sign (PI smaller than or larger than 0.5) of every PI,
sorted by increasing p-value.
}
\examples{
example(fitLMMs, 'spatrans')
writeToXslx(yangPims, "tmpFile.xlsx")
file.remove("tmpFile.xlsx")
}
