% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addCell.R
\name{addCell}
\alias{addCell}
\title{Check in which cell each event lies and add a cell marker}
\usage{
addCell(hypFrame, rois)
}
\arguments{
\item{hypFrame}{The hyperframe}

\item{rois}{the list of list of rois per hyperframe}
}
\value{
the modified hyperframe
}
\description{
Check in which cell each event lies and add a cell marker
}
