% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliaryFunctions.R
\name{addTabObs}
\alias{addTabObs}
\title{Add tables with gene counts to the hyperframe}
\usage{
addTabObs(hypFrame)
}
\arguments{
\item{hypFrame}{The hyperframe}
}
\value{
The hyperframe with tabObs added
}
\description{
Add tables with gene counts to the hyperframe
}
