% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addWeightFunction.R
\name{addWeightFunction}
\alias{addWeightFunction}
\title{Build a weight function based on the data}
\usage{
addWeightFunction(
  hypFrame,
  pis = attr(hypFrame, "pis"),
  designVars,
  maxObs = 1e+05,
  maxFeatures = 500,
  ...
)
}
\arguments{
\item{hypFrame}{The hyperframe with all the data}

\item{pis}{The PIs for which weight functions are constructed}

\item{designVars}{A character vector containing all design factors (both fixed and random), that are also present as variables in hypFrame}

\item{maxObs, maxFeatures}{The maximum number of observations respectively features for fitting the weight function. See details}

\item{...}{Additional arguments passed on to the scam::scam function}
}
\value{
A fitted scam object, which can be used to
}
\description{
Build a weight function based on the data
}
\details{
For computational and memory reasons, for large datasets the trend fitting
is restricted to a subset of the data through the maxObs and maxFeatures parameters.

The scam functions fits a decreasing spline of the variance as a function of the number of observations
}
\examples{
data(Yang)
hypYang = buildHyperFrame(Yang, coordVars = c("x", "y"),
imageVars = c("day", "root", "section"))
#Fit a subset of features to limit computation time
yangPims = estPims(hypYang, pis = c("nn", "nnPair"),
features = attr(hypYang, "features")[1:12])
#Build the weight function for all PIs present
yangObj <- addWeightFunction(yangPims, designVars = c("day", "root"))
}
\seealso{
\link{buildDfMM}
}
