% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buidDfMM.R
\name{buildDfMM}
\alias{buildDfMM}
\title{Build a data frame for a single gene and measure to prepare for mixed model building}
\usage{
buildDfMM(
  resList,
  gene,
  pi = c("nn", "allDist", "nnPair", "allDistPair", "edge", "midpoint", "fixedpoint")
)
}
\arguments{
\item{gene}{A character string indicating the desired gene or gene pair}

\item{pi}{character string indicating the desired pim as outcome in the linear mixed model}

\item{pimRes}{The result of a call to estPims()}

\item{designVars}{Optionally, the design variables to be included in the dataframe. Defaults to all variables in the hyperframe.}
}
\value{
A dataframe
}
\description{
Build a data frame for a single gene and measure to prepare for mixed model building
}
\examples{
data(Yang)
hypYang = buildHyperFrame(Yang, coordVars = c("x", "y"),
designVar = c("day", "root", "section"))
#Fit a subset of features to limit computation time
yangPims = estPims(hypYang, pis = c("nn", "nnPair"),
features = attr(hypYang, "features")[1:12])
#First build the weight function
wf <- buildWeightFunction(yangPims, pi = "nn", hypFrame = hypYang,
designVars = c("day", "root"))
#Now build the data frame for mixed model analysis
dfUniNN = buildDfMM(yangPims, gene = "SmAHK4e", pi  = "nn",
hypFrame = hypYang,  weightFunction = wf)
#A bivariate weight function for the pairs
wfPair = buildWeightFunction(yangPims, pi = "nnPair", hypFrame = hypYang,
designVars = c("day", "root"))
dfBiNN = buildDfMM(yangPims, gene = "SmVND2--SmPINR", pi  = "nnPair",
hypFrame = hypYang,  weightFunction = wfPair, designVars = c("root", "day"))
#Example analysis with linear mixed model
library(lmerTest)
#Use sum coding for day to maintain interpretability of the intercept
mixedMod = lmer(pi - 0.5 ~ day + (1|root), weight = weight, data = dfBiNN,
contrasts = list("day" = "contr.sum"))
summary(mixedMod)
#Evidence for anticorrelation
}
