% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildHyperFrame.R
\name{buildHyperFrame}
\alias{buildHyperFrame}
\alias{buildHyperFrame,data.frame-method}
\alias{buildHyperFrame,matrix-method}
\alias{buildHyperFrame,list-method}
\alias{buildHyperFrame,SpatialExperiment-method}
\title{Built the hyperframe containing all separate point patterns that will
 be used throughout the analysis. Both matrices, dataframe and SpatialExperiment inputs are accepted.}
\usage{
buildHyperFrame(x, ...)

\S4method{buildHyperFrame}{data.frame}(
  x,
  coordVars,
  imageVars,
  coVars = setdiff(names(x), c(imageVars, coordVars)),
  ...
)

\S4method{buildHyperFrame}{matrix}(x, design, covariates, ...)

\S4method{buildHyperFrame}{list}(x, ...)

\S4method{buildHyperFrame}{SpatialExperiment}(x, imageVars, coVars, ...)
}
\arguments{
\item{x}{the input object, see methods("buildHyperFrame")}

\item{...}{additional constructor arguments}

\item{coordVars}{Names of coordinates}

\item{imageVars}{A character vector, together uniquely defining the point pattern (image)}

\item{coVars}{Names of covariates such as gene or cell for each single point}

\item{design}{A single design variable distinguishing the different point patterns}

\item{covariates}{A vector of covariates}

\item{matrix}{The input matrix}

\item{list}{A list of point patterns of class "ppp"}
}
\description{
Built the hyperframe containing all separate point patterns that will
 be used throughout the analysis. Both matrices, dataframe and SpatialExperiment inputs are accepted.
}
\examples{
data(Yang)
hypYang = buildHyperFrame(Yang, coordVars = c("x", "y"),
imageVars = c("day", "root", "section"))
}
