% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildMoransIDataFrame.R
\name{buildMoransIDataFrame}
\alias{buildMoransIDataFrame}
\title{Build a data frame with Moran's I as outcome variable}
\usage{
buildMoransIDataFrame(pi, piMat, weightMats)
}
\arguments{
\item{pi}{character string indicating the desired PI}

\item{piMat}{A data frame. Will be constructed if not provided, for internal use.}

\item{weightMats}{List of weight matrices for Moran's I calculation.}
}
\value{
A modified data frame, containing the estimated Moran's I and its variance
}
\description{
Build a data frame with Moran's I as outcome variable
}
\seealso{
\link[ape]{Moran.I}, \link{buildMoransIWeightMat}
}
