% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildMoransIWeightMat.R
\name{buildMoransIWeightMat}
\alias{buildMoransIWeightMat}
\title{Build a weight matrix based on nearest neighbourship for Moran's I calculations}
\usage{
buildMoransIWeightMat(coordMat, numNNs)
}
\arguments{
\item{coordMat}{A matrix of centroid coordinates}

\item{numNNs}{An integer, the number of nearest neighbours in the
weight matrix for the calculation of the Moran's I statistic}
}
\value{
A sparse matrix of weights for Moran's I statistic, of equal value
for the numNNs nearest neighbours and zero otherwise
}
\description{
Build a weight matrix based on nearest neighbourship for Moran's I calculations
}
\note{
The choice of numNN nearest neighbours is far less memory-glutton than
a weight decaying continuously with distance
}
\seealso{
\link{buildMoransIDataFrame}, \link[ape]{Moran.I}
}
