% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildWeightFunction.R
\name{buildWeightFunction}
\alias{buildWeightFunction}
\title{Build a weight function based on the data}
\usage{
buildWeightFunction(
  pimRes,
  pi = c("nn", "allDist", "nnPair", "allDistPair"),
  hypFrame,
  designVars,
  ...
)
}
\arguments{
\item{pimRes}{The object containing all}

\item{pi}{A character string, the PI required}

\item{hypFrame}{The hyperframe with all the data}

\item{designVars}{A character vector containing all design factors (both fixed and random), that are also present as variables in hypFrame}

\item{...}{Additional arguments passed on to the scam::scam function}
}
\value{
A fitted scam object, which can be used to
}
\description{
Build a weight function based on the data
}
\details{
The scam functions fits a decreasing spline of the variance as a function of the number of observations
}
