% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliaryFunctions.R
\name{buildWeightMat}
\alias{buildWeightMat}
\title{Build a weight matrix for Moran's I calculations}
\usage{
buildWeightMat(coordList)
}
\arguments{
\item{coordList}{A list of coordinate matrices}
}
\value{
A matrix of weights for Moran's I statistic, inversely proportional to the distance between cell centroids
}
\description{
Build a weight matrix for Moran's I calculations
}
