% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcAllDistPI.R
\name{calcAllDistPI}
\alias{calcAllDistPI}
\title{Estimate the probabilistic index for the collection of all distances}
\usage{
calcAllDistPI(pSub, p, ecdfAll, null, ecdfs)
}
\arguments{
\item{pSub}{The subset point pattern with only one feature}

\item{p}{The point pattern}

\item{ecdfAll}{The empirical cumulative distribution function of all distances under the null}

\item{null}{A character vector, indicating how the null distribution is defined. See details.}

\item{ecdfs}{A list of empirical distribution functions, one for each event}
}
\value{
The probabilistic index for all distances
}
\description{
Estimate the probabilistic index for the collection of all distances
}
