% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcBiPIs.R
\name{calcBiPIs}
\alias{calcBiPIs}
\title{Calculate bivaraiet PIs}
\usage{
calcBiPIs(
  p,
  pis,
  null,
  ecdfs,
  nSub,
  ecdfAll,
  features,
  manyPairs,
  verbose,
  allowManyGenePairs
)
}
\arguments{
\item{p}{The point pattern}

\item{pis}{The probabilistic indices to be estimated}

\item{null}{A character vector, indicating how the null distribution is defined. See details.}

\item{ecdfs}{Ecdfs per point}

\item{nSub}{The number of events in the subsampled pattern}

\item{ecdfAll}{Overall ecdf}

\item{features}{A character vector, for which features should the probabilistic indices be calculated?}

\item{manyPairs}{Number of gene pairs to be considered "many"}

\item{verbose}{A boolean, should verbose output be printed}

\item{allowManyGenePairs}{A boolean, set to true to override warning on computation time}
}
\value{
A matrix of bivariate PIs
}
\description{
Calculate bivaraiet PIs
}
