% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcFixedPointDistPI.R
\name{calcFixedPointDistPI}
\alias{calcFixedPointDistPI}
\title{Estimate the PI for the distance to a fixed region of interest}
\usage{
calcFixedPointDistPI(pSub, pointPPP, ecdfAll)
}
\arguments{
\item{pSub}{The subset point pattern containing only a single gene}

\item{ecdfAll}{the cumulative distribution function under the null}

\item{point}{The fixed point to which the distance is to be calculated}
}
\description{
Estimate the PI for the distance to a fixed region of interest
}
