% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcNNpi.R
\name{calcNNPI}
\alias{calcNNPI}
\title{Estimate the PI for the nearest neighbour distances, given a set of ranks}
\usage{
calcNNPI(Ranks, n, m, r = 1)
}
\arguments{
\item{Ranks}{The (approximate) ranks}

\item{n}{the total number of observed distances minus the number of distances
under consideration (the number of black balls in the urn)}

\item{m}{the number of observed distances (white balls in the urn)}

\item{r}{The rank of distances considered, r=1 is nearest neighbour distance}
}
\value{
The estimated probabilistic index
}
\description{
Estimate the PI for the nearest neighbour distances, given a set of ranks
}
\seealso{
\link{pnhyper}
}
