% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcNNpi.R
\name{calcNNPI}
\alias{calcNNPI}
\title{Estimate the PI for the nearest neighbour distances}
\usage{
calcNNPI(pSub, p, null, nSims, rowSortMat, id)
}
\arguments{
\item{pSub}{The subset point pattern containing only a single gene}

\item{p}{The point pattern}

\item{null}{A character vector, indicating how the null distribution is defined. See details.}

\item{nSims}{The number of Monte-Carlo simulations used to determine the null distribution if null = "CSR}
}
\value{
The estimated probabilistic index
}
\description{
Estimate the PI for the nearest neighbour distances
}
