% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcUniPIs.R
\name{calcUniPIs}
\alias{calcUniPIs}
\title{Title}
\usage{
calcUniPIs(
  p,
  pis,
  ecdfsEdgeAndMidpoint,
  owins,
  tabObs,
  null,
  nSims,
  ecdfs,
  nSub,
  ecdfAll,
  idOne
)
}
\arguments{
\item{p}{The point pattern}

\item{pis}{the Probabilitstic indices to be estimated}

\item{ecdfsEdgeAndMidpoint}{ecdfs within the cells}

\item{owins}{The list of windows corresponding to cells}

\item{tabObs}{A table of observed gene frequencies}

\item{null}{A character vector, indicating how the null distribution is defined. See details.}

\item{nSims}{The number of Monte-Carlo simulations used to determine the null distribution if null = "CSR}

\item{ecdfs}{Ecdfs per point}

\item{ecdfAll}{Overall ecdf}
}
\value{
PIs for every feature
}
\description{
Title
}
