% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcUniPIs.R
\name{calcUniPIs}
\alias{calcUniPIs}
\title{Calculate univariate PIs for all features}
\usage{
calcUniPIs(
  p,
  pis,
  verbose,
  ecdfsEdgeAndMidpoint,
  owins,
  tabObs,
  null,
  nSims,
  ecdfs,
  nSub,
  ecdfAll,
  idOne,
  features,
  ecdfFixedPoint,
  pointPPP
)
}
\arguments{
\item{p}{The point pattern}

\item{pis}{the Probabilitstic indices to be estimated}

\item{verbose}{A boolean, should verbose output be printed}

\item{ecdfsEdgeAndMidpoint}{ecdfs within the cells}

\item{owins}{The list of windows corresponding to cells}

\item{tabObs}{A table of observed gene frequencies}

\item{null}{A character vector, indicating how the null distribution is defined. See details.}

\item{nSims}{The number of Monte-Carlo simulations used to determine the null distribution if null = "CSR}

\item{ecdfs}{Ecdfs per point}

\item{nSub}{The number of events in the subsampled pattern}

\item{ecdfAll}{Overall ecdf}

\item{idOne}{Index of features with only one observation, to be excluded}

\item{features}{A character vector, for which features should the probablilistic indices be calculated?}

\item{ecdfFixedPoint}{The ecdf of the fixed point}

\item{pointPPP}{the point pattern containing the fixed point}
}
\value{
PIs for every feature
}
\description{
Calculate univariate PIs for all features
}
