% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcWindowDistPI.R
\name{calcWindowDistPI}
\alias{calcWindowDistPI}
\title{Estimate the PI for the distance to a fixed region of interest}
\usage{
calcWindowDistPI(pSub, owins, ecdfAll, midPoint = FALSE)
}
\arguments{
\item{pSub}{The subset point pattern containing only a single gene}

\item{ecdfAll}{the cumulative distribution function under the null}

\item{edge}{The psp object to which the distances are calculated}
}
\value{
The estimated probabilistic index
}
\description{
Estimate the PI for the distance to a fixed region of interest
}
