% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcWindowDistPI.R
\name{calcWindowDistPI}
\alias{calcWindowDistPI}
\title{Estimate the PI for the distance to a fixed region of interest}
\usage{
calcWindowDistPI(pSub, owins, ecdfAll, towhat)
}
\arguments{
\item{pSub}{The subset point pattern containing only a single gene}

\item{owins}{the list of windows describing the cells}

\item{ecdfAll}{the cumulative distribution function under the null}

\item{towhat}{a character vector, either "edge" or "midpoint", indicating with respect to what the distance should be calculated}
}
\value{
The estimated probabilistic index
}
\description{
Estimate the PI for the distance to a fixed region of interest
}
\details{
Analysis of the distance to the border was introduced by \insertCite{Joyner2013}{spatrans} in the form of the B-function.
The independent evaluations of the B-functions per cell are here returned as realizations of the probabilistic index.
}
\references{
\insertAllCited{}
}
\seealso{
\link{addCell}
}
