% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{centerNumeric}
\alias{centerNumeric}
\title{Center numeric variables}
\usage{
centerNumeric(x)
}
\arguments{
\item{x}{The dataframe whose numeric variables are being centered}
}
\value{
The adapted dataframe
}
\description{
Center numeric variables
}
\examples{
df = data.frame(a = rnorm(10), b = sample(c(TRUE, FALSE), 10, replace = TRUE))
dfCen = centerNumeric(df)
mean(dfCen$a)
}
