% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkAttr.R
\name{checkAttr}
\alias{checkAttr}
\title{Check if the required pi's are present in theobject}
\usage{
checkAttr(x, pi)
}
\arguments{
\item{x}{The result of the PI calculation, or a weight function}

\item{pi}{A character string indicating the desired PI}
}
\value{
Throws an error when the PIs are not found, otherwise returns invisible
}
\description{
Check if the required pi's are present in theobject
}
