% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructDesignVars.R
\name{constructDesignVars}
\alias{constructDesignVars}
\title{Run checks on design variables, or construct them if missing}
\usage{
constructDesignVars(designVars, lowestLevelVar, allCell, resList)
}
\arguments{
\item{designVars}{The initial design variables}

\item{lowestLevelVar}{Variable indicating the lowest level of nesting}

\item{allCell}{A boolean, are all PIs cell-related?}

\item{resList}{The results list}
}
\value{
A vector of design variables
}
\description{
Run checks on design variables, or construct them if missing
}
\seealso{
\link{buildDataFrame}
}
