% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossdistWrapper.R
\name{crossdistWrapper}
\alias{crossdistWrapper}
\title{A wrapper for C-functions calculating cross-distance matrix fast and with
low memory usage. Big matrices are also supported}
\usage{
crossdistWrapper(x, y, returnBigMatrix = FALSE, tmpFile = NULL)
}
\arguments{
\item{x, y}{the matrices or point patterns between which to calculate the
cross distances}

\item{returnBigMatrix}{a boolean, should the result be returned as
a big matrix (low RAM use) or a regular r-matrix (faster)}
}
\value{
a matrix of cross distances
}
\description{
A wrapper for C-functions calculating cross-distance matrix fast and with
low memory usage. Big matrices are also supported
}
