% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estGradients.R
\name{estGradients}
\alias{estGradients}
\title{Estimate gradients for single-molecule localization patterns}
\usage{
estGradients(
  hypFrame,
  gradients = c("overall", if (!is.null(hypFrame$owins)) "cell"),
  fixedEffects = NULL,
  verbose = FALSE,
  features = getFeatures(hypFrame),
  silent = TRUE,
  loopFun = "bplapply",
  ...
)
}
\arguments{
\item{hypFrame}{the hyperframe}

\item{gradients}{The gradients indices to be estimated}

\item{fixedEffects}{Character vector of fixed effects present in the hyperframe,
modifying the baseline intensity. See details.}

\item{verbose}{A boolean, whether to report on progress of the fitting process.}

\item{features}{A character vector, for which features should the
gradients indices be calculated?}

\item{silent}{A bolean, should error messages from ppm.ppp be printed?}

\item{loopFun}{The function to use to loop over the features.
Defaults to bplapply except when looping over features within cells}

\item{...}{additional arguments, passed on to \link{fitGradient}.}
}
\value{
A list with the hyperframe and the estimated gradients
}
\description{
Estimate gradients on all point patterns of a hyperframe,
and integrate the results in the same hyperframe
}
\details{
The fixed effects modify the baseline intensity of the point pattern, not the gradient!
}
\examples{
data(Yang)
hypYang <- buildHyperFrame(Yang,
    coordVars = c('x', 'y'),
    imageVars = c('day', 'root', 'section')
)
yangGrads <- estGradients(hypYang, features = getFeatures(hypYang)[seq_len(5)])
Overall Gradients
}
\seealso{
\link{fitGradient}, \link{estGradientsSingle}
}
