% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estGradients.R
\name{estGradients}
\alias{estGradients}
\title{Estimate gradients for single-molecule localization patterns}
\usage{
estGradients(
  hypFrame,
  gradients = c("overall", if (!is.null(hypFrame$owins)) "cell"),
  verbose = TRUE,
  features = getFeatures(hypFrame),
  silent = TRUE,
  ...
)
}
\arguments{
\item{hypFrame}{the hyperframe}

\item{gradients}{The gradients indices to be estimated}

\item{verbose}{A boolean, whether to report on progress of the fitting process.}

\item{features}{A character vector, for which features should the
gradients indices be calculated?}

\item{silent}{A bolean, should error messages from ppm.ppp be printed?}

\item{...}{additional arguments, passed on to \link{fitGradient}.}
}
\value{
The hyperframe with the estimated gradients present in it
}
\description{
Estimate gradients on all point patterns of a hyperframe,
and integrate the results in the same hyperframe
}
\examples{
data(Yang)
hypYang <- buildHyperFrame(Yang,
    coordVars = c('x', 'y'),
    imageVars = c('day', 'root', 'section')
)
yangGrads <- estGradients(hypYang)
Overall Gradients
}
\seealso{
\link{fitGradient, estGradientsSingle}
}
