% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estGradients.R
\name{estGradientsSingle}
\alias{estGradientsSingle}
\title{A wrapper function for the estimation of the different gradients, applied to individual point patterns}
\usage{
estGradientsSingle(hypFrame, gradients, fixedEffects, owins, ...)
}
\arguments{
\item{hypFrame}{the hyperframe}

\item{gradients}{The gradients indices to be estimated}

\item{...}{Passed onto fitGradient}
}
\value{
A list of data frames with estimated gradients per gene
\item{overall}{overall gradients}
\item{cell}{gradients per cell}
}
\description{
A wrapper function for the estimation of the different gradients, applied to individual point patterns
}
\seealso{
\link{estGradients}
}
