% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estGradients.R
\name{estGradientsSingle}
\alias{estGradientsSingle}
\title{A wrapper function for the estimation of the different gradients, applied to individual point patterns}
\usage{
estGradientsSingle(
  p,
  tabObs,
  gradients,
  owins = NULL,
  loopFun = "bplapply",
  features,
  ...
)
}
\arguments{
\item{p}{The point pattern}

\item{tabObs}{A table of observed gene frequencies}

\item{loopFun}{The function to use to loop over the features.
Defaults to bplapply except when looping over features within cells}

\item{features}{A character vector, for which features should the
probabilistic indices be calculated?}

\item{...}{Passed onto fitGradient}
}
\value{
A list of data frames with estimated gradients per gene
\item{overall}{overall gradients}
\item{cell}{gradients per cell}
}
\description{
A wrapper function for the estimation of the different gradients, applied to individual point patterns
}
\seealso{
\link{estGradients}
}
