% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estPims.R
\name{estPims}
\alias{estPims}
\title{Estimate pims on a hyperframe}
\usage{
estPims(hypFrame, ...)
}
\arguments{
\item{hypFrame}{the hyperframe}
}
\value{
A list of estimated pims
}
\description{
Estimate pims on a hyperframe
}
