% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estPims.R
\name{estPims}
\alias{estPims}
\title{Estimate pims on a hyperframe}
\usage{
estPims(
  hypFrame,
  pis = c("nn", "allDist", "nnPair", "allDistPair", "edge", "midpoint", "fixedpoint"),
  null = c("background", "CSR"),
  features = attr(hypFrame, "features"),
  ...
)
}
\arguments{
\item{hypFrame}{the hyperframe}

\item{pis}{the Probabilitstic indices to be estimated}

\item{null}{A character vector, indicating how the null distribution is defined. See details.}

\item{features}{A character vector, for which features should the probablilistic indices be calculated?}

\item{...}{additional arguments, passed on to estPimsSingle}
}
\value{
A list of estimated pims
}
\description{
Estimate pims on a hyperframe
}
\examples{
data(Yang)
hypYang = suppressWarnings(buildHyperFrame(Yang, coordVars = c("x", "y"), designVar = c("day", "root", "section")))
yangPims = estPims(hypYang, pis = c("nn", "nnPair"))
}
