% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estPims.R
\name{estPimsSingle}
\alias{estPimsSingle}
\title{A wrapper function for the different probabilistic indices (PIs)}
\usage{
estPimsSingle(
  p,
  pis,
  null,
  tabObs,
  owins = NULL,
  centroids = NULL,
  window = p$window,
  loopFun = "bplapply"
)
}
\arguments{
\item{p}{The point pattern}

\item{pis}{The probabilistic indices to be estimated}

\item{null}{A character vector, indicating how the null distribution is
defined. See details.}

\item{tabObs}{A table of observed gene frequencies}

\item{owins, centroids}{The list of windows corresponding to cells,
and their centroids}

\item{window}{An window of class owin, in which events can occur}

\item{loopFun}{The function to use to loop over the features.
Defaults to bplapply except when looping over features within cells}
}
\value{
A list of data frames with estimated PIs per gene and/or gene pair:
\item{pointDists}{PIs for pointwise distances overall}
\item{windowDists}{PIs for distances to cell wall or centroid}
\item{withinCellDists}{PIs for pointwise distances within cell}
}
\description{
A wrapper function for the different probabilistic indices (PIs)
}
