% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estPims.R
\name{estPimsSingle}
\alias{estPimsSingle}
\title{A wrapper function for the different probabilistic indices (PIs)}
\usage{
estPimsSingle(
  p,
  pis,
  null,
  tabObs,
  nPointsAll = switch(null, background = 10000, CSR = 1000),
  nPointsAllWin = 2000,
  features = NULL,
  owins = NULL,
  centroids = NULL,
  window = p$window
)
}
\arguments{
\item{p}{The point pattern}

\item{pis}{The probabilistic indices to be estimated}

\item{null}{A character vector, indicating how the null distribution is
defined. See details.}

\item{tabObs}{A table of observed gene frequencies}

\item{nPointsAll}{How many points to subsample or simulate to calculate
overall interpoint distance and distance to point.
This parameter is a strong driver of memory and cpu usage.}

\item{nPointsAllWin}{How many points to subsample or simulate
to calculate distance to cell edge or midpoint distribution}

\item{features}{A character vector, for which features should the
probabilistic indices be calculated?}

\item{owins, centroids}{The list of windows corresponding to cells,
and their centroids}

\item{window}{An window of class owin, in which events can occur}
}
\value{
Data frames with estimated quantities per gene and/or gene pair
}
\description{
A wrapper function for the different probabilistic indices (PIs)
}
