% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estPims.R
\name{estPimsSingle}
\alias{estPimsSingle}
\title{A wrapper function for the different probabilistic indices (PIs)}
\usage{
estPimsSingle(
  p,
  pis,
  null,
  tabObs,
  nPointsAll = switch(null, background = 1e+05, CSR = 1000),
  nPointsAllWithinCell = switch(null, background = 10000, CSR = 1000),
  nPointsAllWin = 2000,
  features = NULL,
  owins = NULL,
  centroids = NULL,
  window = p$window,
  loopFun = "bplapply",
  minDiff = 20
)
}
\arguments{
\item{p}{The point pattern}

\item{pis}{The probabilistic indices to be estimated}

\item{null}{A character vector, indicating how the null distribution is
defined. See details.}

\item{tabObs}{A table of observed gene frequencies}

\item{nPointsAll, nPointsAllWithinCell}{How many points to subsample or simulate to calculate
overall interpoint distance and distance to point.
This parameter is a strong driver of computation time.
The second argument applies to within cell calculations.}

\item{nPointsAllWin}{How many points to subsample or simulate
to calculate distance to cell edge or midpoint distribution}

\item{features}{A character vector, for which features should the
probabilistic indices be calculated?}

\item{owins, centroids}{The list of windows corresponding to cells,
and their centroids}

\item{window}{An window of class owin, in which events can occur}

\item{loopFun}{The function to use to loop over the features.
Defaults to bplapply except when looping over features within cells}

\item{minDiff}{An integer, the minimum number of events from other genes
needed for calculation of background PIs}
}
\value{
Data frames with estimated quantities per gene and/or gene pair
}
\description{
A wrapper function for the different probabilistic indices (PIs)
}
