% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estPis.R
\name{estPis}
\alias{estPis}
\title{Estimate probabilistic indices for (co)localization on a hyperframe}
\usage{
estPis(
  hypFrame,
  pis = c("nn", "nnPair", "edge", "centroid", "nnCell", "nnPairCell"),
  verbose = TRUE,
  null = c("background", "CSR"),
  nPointsAll = switch(null, background = 10000, CSR = 1000),
  nPointsAllWithinCell = switch(null, background = 10000, CSR = 500),
  nPointsAllWin = 1000,
  minDiff = 20,
  features = getFeatures(hypFrame),
  ...
)
}
\arguments{
\item{hypFrame}{the hyperframe}

\item{pis}{The probabilistic indices to be estimated}

\item{verbose}{a boolean, whether to report on progress of fitting}

\item{null}{A character vector, indicating how the null distribution is
defined. See details.}

\item{nPointsAll, nPointsAllWithinCell}{How many points to subsample or simulate to calculate
overall interpoint distance and distance to point.
This parameter is a strong driver of computation time.
The second argument applies to within cell calculations.}

\item{nPointsAllWin}{How many points to subsample or simulate
to calculate distance to cell edge or centroid distribution}

\item{minDiff}{An integer, the minimum number of events from other genes
needed for calculation of background PIs}

\item{features}{A character vector, for which features should the
probabilistic indices be calculated?}

\item{...}{additional arguments, passed on to estPisSingle}
}
\value{
The hyperframe with the estimated PIMs present in it
}
\description{
Estimate different probabilistic indices for (co)localization
on  a hyperframe, and integrate the results in the same hyperframe
}
\details{
The null distribution used to calculate the PIs. Can be either 'background',
in which case the observed distributions of all genes is used. Alternatively,
for null = 'CSR', Monte-Carlo simulation under complete spatial randomness
 is performed within the given window.

The 'nn' prefix indicates that nearest neighbour distances are being used,
whereas 'all' indicates all distances are being used.
The suffix 'Pair' indicates that bivariate probabilistic indices,
testing for co- and antilocalization are being used.
'edge' and 'centroid' calculate the distance to the edge respectively
 the centroid of the windows added using the addCell() function.
The suffix 'Cell' indicates distances are being calculated within cells only.
}
\examples{
data(Yang)
hypYang <- buildHyperFrame(Yang,
    coordVars = c("x", "y"),
    imageVars = c("day", "root", "section")
)
# Fit a subset of features to limit computation time
yangPims <- estPis(hypYang[c(seq_len(5), seq(25, 29)), ], pis = "nn")
# Univariate nearest neighbour distances
}
