% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalWeightFun.R
\name{evalWeightFunction}
\alias{evalWeightFunction}
\title{Evaluate the weight function and return (unnormalized) weights}
\usage{
evalWeightFunction(wf, newdata)
}
\arguments{
\item{wf}{The weight function}

\item{newdata}{Optional, a data frame with new data}
}
\value{
A vector of predictions
}
\description{
Evaluate the weight function and return (unnormalized) weights
}
\examples{
data(Yang)
hypYang = buildHyperFrame(Yang, coordVars = c("x", "y"),
designVar = c("day", "root", "section"))
#Fit a subset of features to limit computation time
yangPims = estPims(hypYang, pis = c("nn", "nnPair"),
features = attr(hypYang, "features")[1:20])
#First build the weight function
wf <- buildWeightFunction(yangPims, pi = "nn", hypFrame = hypYang,
designVars = c("day", "root"))
evalWeightFunction(wf, newdata = data.frame("NP" = 1))
}
