% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildWeightFunction.R
\name{evalWeightFunction}
\alias{evalWeightFunction}
\title{Evaluate the weight function and return (unnormalized) weights}
\usage{
evalWeightFunction(wf, newdata = NULL)
}
\arguments{
\item{wf}{The weight function}

\item{newdata}{Optional, a data frame with new data}
}
\value{
A vector of predictions
}
\description{
Evaluate the weight function and return (unnormalized) weights
}
