% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalWeightFun.R
\name{evalWeightFunction}
\alias{evalWeightFunction}
\title{Evaluate the weighting function and return (unnormalized) weights}
\usage{
evalWeightFunction(wf, newdata)
}
\arguments{
\item{wf}{The weighting function}

\item{newdata}{Optional, a data frame with new data}
}
\value{
A vector of predictions
}
\description{
Evaluate the weighting function and return (unnormalized) weights
}
