% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalWeightFun.R
\name{evalWeightFunction}
\alias{evalWeightFunction}
\title{Evaluate a variance weighting function}
\usage{
evalWeightFunction(wf, newdata)
}
\arguments{
\item{wf}{The weighting function}

\item{newdata}{Optional, a data frame with new data}
}
\value{
A vector of predictions
}
\description{
Evaluate the variance weighting function to  return unnormalized weights
}
\examples{
data(Yang)
hypYang <- buildHyperFrame(Yang,
    coordVars = c("x", "y"),
    imageVars = c("day", "root", "section")
)
yangPims <- estPis(hypYang[c(seq_len(5), seq(25, 29)), ],
    nPointsAll = 1e3,
    pis = "nn", features = getFeatures(hypYang)[1:10]
)
# First Build the weighting function
yangObj <- addWeightFunction(yangPims, designVars = c("day", "root"))
evalWeightFunction(yangObj$Wfs$nn, newdata = data.frame("NP" = 2))
}
