% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalWeightFun.R
\name{evalWeightFunction}
\alias{evalWeightFunction}
\title{Evaluate the weighting function and return (unnormalized) weights}
\usage{
evalWeightFunction(wf, newdata)
}
\arguments{
\item{wf}{The weighting function}

\item{newdata}{Optional, a data frame with new data}
}
\value{
A vector of predictions
}
\description{
Evaluate the weighting function and return (unnormalized) weights
}
\examples{
data(Yang)
hypYang <- buildHyperFrame(Yang, coordVars = c('x', 'y'),
    imageVars = c('day', 'root', 'section')
)
yangPims <- estPims(hypYang[c(seq_len(5), seq(25, 29)),],
    pis = 'nn', features = getFeatures(hypYang)[1:10])
# First Build the weighting function
yangObj <- addWeightFunction(yangPims, designVars = c('day', 'root'))
evalWeightFunction(yangObj$Wfs$nn, newdata = data.frame('NP' = 2))
}
