% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractResults.R
\name{extractResults}
\alias{extractResults}
\title{Extract results from a list of fitted LMMs}
\usage{
extractResults(
  models,
  hypFrame,
  subSet = "piMod",
  fixedVars = NULL,
  method = "BH",
  increment = switch(subSet, piMod = 0.5, moranMod = 0)
)
}
\arguments{
\item{models}{The models}

\item{hypFrame}{The original hyperframe}

\item{fixedVars}{The fixed effects for which the effect is to be reported}

\item{method}{passed onto p.adjust}
}
\value{
A list of matrices, all containing estimate, standard error,
p-value and ajdusted p-value
}
\description{
Extract results from a list of fitted LMMs
}
\seealso{
\link{fitLMMs}
}
