% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractResults.R
\name{extractResults}
\alias{extractResults}
\title{Extract results from a list of fitted LMMs. For internal use mainly.}
\usage{
extractResults(
  models,
  hypFrame,
  subSet = "piMod",
  fixedVars = NULL,
  method = "BH"
)
}
\arguments{
\item{models}{The models}

\item{hypFrame}{The original hyperframe}

\item{subSet}{The name of the subset to be extracted, either PI or Moran's I}

\item{fixedVars}{The fixed effects for which the effect is to be reported}

\item{method}{Multiplicity correction method passed onto p.adjust}
}
\value{
A list of matrices, all containing estimate, standard error,
p-value and ajdusted p-value
}
\description{
Extract results from a list of fitted LMMs. For internal use mainly.
}
\seealso{
\link{fitLMMs}, \link{p.adjust}
}
