% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findOverlap.R
\name{findOverlap}
\alias{findOverlap}
\title{Try to find overlap between list of windows, and throw error when found}
\usage{
findOverlap(owins, returnIds = FALSE)
}
\arguments{
\item{owins}{the list of windows}

\item{returnIds}{A boolean, should the indices of the overlap be returned?
If FALSE an error is thrown at the first overlap}
}
\value{
Throws an error when overlap found, otherwise returns invisible
}
\description{
Try to find overlap between list of windows, and throw error when found
}
\examples{
library(spatstat.geom)
owins = replicate(10, owin(xrange = runif(1)+c(0,0.2),
yrange = runif(1)+c(0,0.1)), simplify = FALSE)
idOverlap = findOverlap(owins, returnIds = TRUE)
}
