% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitGradient.R
\name{fitGradient}
\alias{fitGradient}
\title{Fit a gradient to a point pattern}
\usage{
fitGradient(hypFrame, fixedEffects = NULL, returnModel = FALSE, silent, ...)
}
\arguments{
\item{hypFrame}{the hyperframe}

\item{returnModel}{A boolean, should the entire model be returned?
Otherwise the p-value is returned}

\item{...}{passed onto spatstat.model::ppm}
}
\value{
A vector of of length 3 with p-value and variances of the random effects,
or a mppm model when returnModel is true
}
\description{
A Poisson process is fitted to the data assuming exponential relationship
between x and y variables and intensity
}
\seealso{
\link{estGradients}
}
