% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitLMMs.R
\name{fitLMMs}
\alias{fitLMMs}
\title{Title}
\usage{
fitLMMs(
  pimRes,
  pi = c("nn", "allDist", "nnPair", "allDistPair", "edge", "midpoint", "fixedpoint"),
  weightFunction,
  hypFrame,
  fixedVars,
  randomVars
)
}
\arguments{
\item{pimRes}{The result of a call to estPims()}

\item{pi}{character string indicating the desired pim as outcome in the linear mixed model}

\item{weightFunction}{The weighing function, obtained by the buildWeightFunction() function}

\item{hypFrame}{The hyperframe with the data}
}
\value{
A fitted linear mixed model of class 'lmerTest'
}
\description{
Title
}
\examples{
#' data(Yang)
hypYang = buildHyperFrame(Yang, coordVars = c("x", "y"),
designVar = c("day", "root", "section"))
yangPims = estPims(hypYang, pis = "nn")
#First build the weight function
wf <- buildWeightFunction(yangPims, pi = "nn", hypFrame = hypYang,
designVars = c("day", "root"))
fittedModels = fitLMMs(yangPims, pi = "nn", weightFunction = wf, fiexedVars = "time", randomVars = "root")
}
