% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitLMMs.R
\name{fitLMMs}
\alias{fitLMMs}
\title{Fit linear mixed nodels for all features of a pimRes object}
\usage{
fitLMMs(
  resList,
  pi = c("nn", "allDist", "nnPair", "allDistPair", "edge", "midpoint", "fixedpoint"),
  fixedVars,
  randomVars,
  verbose = TRUE
)
}
\arguments{
\item{pi}{character string indicating the desired pim as outcome in the linear mixed model}

\item{fixedVars, randomVars}{Names of fixed and random variables}
}
\value{
A fitted linear mixed model of class 'lmerTest'
}
\description{
Fit linear mixed nodels for all features of a pimRes object
}
\examples{
data(Yang)
hypYang = buildHyperFrame(Yang, coordVars = c("x", "y"),
imageVars = c("day", "root", "section"))
#Fit a subset of features to limit computation time
yangPims = estPims(hypYang, pis = "nn", features = attr(hypYang, "features")[seq_len(20)])
#First build the weight function
yangObj <- addWeightFunction(yangPims, designVars = c("day", "root"))
fittedModels = fitLMMs(yangObj, fixedVars = "day", randomVars = "root")
}
