% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitLMMs.R
\name{fitLMMs}
\alias{fitLMMs}
\title{Fit linear (mixed) models fitting for all probabilistic indices (PIs)}
\usage{
fitLMMs(
  obj,
  pis = obj$pis,
  fixedVars = NULL,
  randomVars = NULL,
  verbose = TRUE,
  returnModels = FALSE,
  Formula = NULL,
  randomNested = TRUE,
  features = getFeatures(obj),
  ...
)
}
\arguments{
\item{obj}{The result object}

\item{pis}{Optional, the pis required. Defaults to all pis in the object}

\item{verbose}{a boolean, should output be printed?}

\item{...}{Passed onto fitLMMsSingle}
}
\value{
A list of fitted objects
}
\description{
The PI is used as outcome variable in a linear (mixed) model,
with design variables as regressors.
}
\examples{
example(addWeightFunction, "spatrans")
lmmModels <- fitLMMs(yangObj, fixedVars = "day", randomVars = "root")
}
