% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitLMMs.R
\name{fitLMMs}
\alias{fitLMMs}
\title{A wrapper function for linear mixed model fitting for all PIs}
\usage{
fitLMMs(obj, pis = obj$pis, verbose = TRUE, ...)
}
\arguments{
\item{obj}{The result object}

\item{pis}{Optional, the pis required. Defaults to all pis in the object}

\item{verbose}{a boolean, should output be printed?}

\item{...}{Passed onto fitLMMsSingle}
}
\value{
A list of fitted objects
}
\description{
A wrapper function for linear mixed model fitting for all PIs
}
\examples{
data(Yang)
hypYang <- buildHyperFrame(Yang,
    coordVars = c('x', 'y'),
    imageVars = c('day', 'root', 'section')
)
# Fit a subset of features to limit computation time
yangPims <- estPims(hypYang[c(seq_len(5), seq(25, 29)),], pis = c('nn', 'nnPair'))
# First build the weighting function
yangPims <- addWeightFunction(yangPims, designVars = c('day', 'root'))
lmmModels <- fitLMMs(yangPims, fixedVars = 'day', randomVars = 'root')
}
