% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitLMMs.R
\name{fitLMMs}
\alias{fitLMMs}
\title{Fit linear (mixed) models fitting for all probabilistic indices (PIs)}
\usage{
fitLMMs(
  obj,
  pis = obj$pis,
  fixedVars = NULL,
  randomVars = NULL,
  verbose = TRUE,
  returnModels = FALSE,
  Formula = NULL,
  randomNested = TRUE,
  features = getFeatures(obj),
  moranFormula = NULL,
  addMoransI = FALSE,
  ...
)
}
\arguments{
\item{obj}{The result object}

\item{pis}{Optional, the pis required. Defaults to all pis in the object}

\item{fixedVars}{Names of fixed effects}

\item{randomVars}{Names of random variables, possibly in a single vector to
reflect nesting structure, see details.}

\item{verbose}{A boolean, should the formula be printed?}

\item{returnModels}{a boolean: should the full models be returned?
Otherwise only summary statistics are returned}

\item{Formula}{A formula; if not supplied it will be constructed
from the fixed and random variables}

\item{randomNested}{A boolean, indicating if random effects are nested within
point patterns. See details.}

\item{features}{The features for which to fit linear mixed models.
Defaults to all features in the object}

\item{moranFormula}{Formula for Moran's I model fitting}

\item{addMoransI}{A boolean, include Moran's I in the calculation}

\item{...}{Passed onto fitLMMsSingle}
}
\value{
A list of fitted objects
}
\description{
The PI is used as outcome variable in a linear (mixed) model,
with design variables as regressors.
}
\examples{
example(addWeightFunction, "spatrans")
lmmModels <- fitLMMs(yangObj, fixedVars = "day", randomVars = "root")
}
