% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitLMMs.R
\name{fitLMMs}
\alias{fitLMMs}
\title{Fti linear mixed nodels for all features of a pimRes object}
\usage{
fitLMMs(
  pimRes,
  pi = c("nn", "allDist", "nnPair", "allDistPair", "edge", "midpoint", "fixedpoint"),
  weightFunction,
  hypFrame,
  fixedVars,
  randomVars
)
}
\arguments{
\item{pimRes}{The result of a call to estPims()}

\item{pi}{character string indicating the desired pim as outcome in the linear mixed model}

\item{weightFunction}{The weighing function, obtained by the buildWeightFunction() function}

\item{hypFrame}{The hyperframe with the data}
}
\value{
A fitted linear mixed model of class 'lmerTest'
}
\description{
Fti linear mixed nodels for all features of a pimRes object
}
