% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitLMMs.R
\name{fitLMMs}
\alias{fitLMMs}
\title{Fit linear mixed nodels for all features of a pimRes object}
\usage{
fitLMMs(
  resList,
  pi = c("nn", "allDist", "nnPair", "allDistPair", "edge", "midpoint"),
  fixedVars,
  randomVars,
  verbose = TRUE,
  resultsOnly = TRUE,
  Formula = NULL
)
}
\arguments{
\item{resList}{The result of a call to estPims()}

\item{pi}{character string indicating the desired pim as outcome
in the linear mixed model}

\item{fixedVars, randomVars}{Names of fixed and random variables}

\item{verbose}{A boolean, should the formula be printed?}

\item{resultsOnly}{a boolean: should only the results be returned?
If false, also the linear mixed models are returned}

\item{Formula}{A formula; if not supplied it will be constructed
from the fixed and random variables}
}
\value{
A fitted linear mixed model of class 'lmerTest'
}
\description{
Fit linear mixed nodels for all features of a pimRes object
}
\examples{
data(Yang)
hypYang <- buildHyperFrame(Yang,
    coordVars = c("x", "y"),
    imageVars = c("day", "root", "section")
)
# Fit a subset of features to limit computation time
yangPims <- estPims(hypYang, pis = "nn", features = attr(hypYang, "features")[seq_len(20)])
# First build the weight function
yangObj <- addWeightFunction(yangPims, designVars = c("day", "root"))
fittedModels <- fitLMMs(yangObj, fixedVars = "day", randomVars = "root")
}
