% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitLMMs.R
\name{fitLMMs}
\alias{fitLMMs}
\title{Fit linear mixed nodels for all features of a pimRes object}
\usage{
fitLMMs(
  resList,
  pi,
  fixedVars = NULL,
  randomVars = NULL,
  verbose = TRUE,
  returnModels = FALSE,
  Formula = NULL,
  randomNested = TRUE
)
}
\arguments{
\item{resList}{The result of a call to estPims()}

\item{pi}{character string indicating the desired pim as outcome
in the linear mixed model}

\item{fixedVars}{Names of fixed effects}

\item{randomVars}{Names of random variables, possibly in a single vector to
reflect nesting structure, see details.}

\item{verbose}{A boolean, should the formula be printed?}

\item{returnModels}{a boolean: should the full models be returned?
Otherwise only summary statistics are returned}

\item{Formula}{A formula; if not supplied it will be constructed
from the fixed and random variables}

\item{randomNested}{A boolean, indicating if random effects are nested within
point patterns. See details.}
}
\value{
A fitted linear mixed model of class 'lmerTest'
}
\description{
Fit linear mixed nodels for all features of a pimRes object
}
\details{
Genes or gene pairs with insufficient observations will be silently
omitted. When randomVars is provided as a vector, independent random
intercepts are fitted for them by default. Providing them separated by "\" or
":" as in the lmer formulas is also allowed to reflect nesting structure.

It is by default assumed that random effects are nested within the point
 patterns. This means for instance that cells with the same name but from
 different point patterns are assigned to different random effects. Set
 "randomNested" to FALSE to override this behaviour.
}
\examples{
data(Yang)
hypYang <- buildHyperFrame(Yang,
    coordVars = c("x", "y"),
    imageVars = c("day", "root", "section")
)
# Fit a subset of features to limit computation time
yangPims <- estPims(hypYang[c(seq_len(5), seq(25, 29)),], pis = "nn",
    features = attr(hypYang, "features")[seq_len(10)]
)
# First build the weighting function
yangObj <- addWeightFunction(yangPims, designVars = c("day", "root"))
lmmModels <- fitLMMs(yangObj, fixedVars = "day", randomVars = "root",
    pi = "nn"
)
}
\seealso{
\link{buildDfMM},\link{getResults}
#TO DO: write wrapper over all PIs!
}
