% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitLMMs.R
\name{fitLMMs}
\alias{fitLMMs}
\title{Fit linear mixed nodels for all features of a pimRes object}
\usage{
fitLMMs(
  resList,
  pi,
  fixedVars = NULL,
  randomVars = NULL,
  verbose = TRUE,
  returnModels = FALSE,
  Formula = NULL
)
}
\arguments{
\item{resList}{The result of a call to estPims()}

\item{pi}{character string indicating the desired pim as outcome
in the linear mixed model}

\item{fixedVars, randomVars}{Names of fixed and random variables}

\item{verbose}{A boolean, should the formula be printed?}

\item{returnModels}{a boolean: should the full models be returned?
Otherzise only summary statistics are returned}

\item{Formula}{A formula; if not supplied it will be constructed
from the fixed and random variables}
}
\value{
A fitted linear mixed model of class 'lmerTest'
}
\description{
Fit linear mixed nodels for all features of a pimRes object
}
\examples{
data(Yang)
hypYang <- buildHyperFrame(Yang,
    coordVars = c("x", "y"),
    imageVars = c("day", "root", "section")
)
# Fit a subset of features to limit computation time
yangPims <- estPims(hypYang, pis = "nn",
features = attr(hypYang, "features")[seq_len(20)])
# First build the weighting function
yangObj <- addWeightFunction(yangPims, designVars = c("day", "root"))
lmmModels <- fitLMMs(yangObj, fixedVars = "day", randomVars = "root",
pi = "nn")
}
\seealso{
\link{buildDfMM},\link{getResults}
#FIX ME: write wrapper over all PIs!
}
