% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitLMMs.R
\name{fitLMMsAll}
\alias{fitLMMsAll}
\title{A wrapper function for linear mixed model fitting for all PIs}
\usage{
fitLMMsAll(obj, pis = obj$pis, ...)
}
\arguments{
\item{obj}{The result object}

\item{pis}{Optional, the pis required. Defaults to all pis in the object}

\item{...}{Passed onto fitLMMs}
}
\value{
A list of fitted objects
}
\description{
A wrapper function for linear mixed model fitting for all PIs
}
\examples{
data(Yang)
hypYang <- buildHyperFrame(Yang, coordVars = c('x', 'y'),
imageVars = c('day', 'root', 'section'))
yangPims <- estPims(hypYang[c(seq_len(5), seq(25, 29)),],
features = getFeatures(hypYang)[seq_len(5)],
pis = c('nn', 'nnPair'))
yangPims <- addWeightFunction(yangPims, designVars = c('day', 'root'))
allModels = fitLMMsAll(yangPims, fixedVars = 'day', randomVars = 'root')
}
