% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitLMMs.R
\name{fitLMMsSingle}
\alias{fitLMMsSingle}
\title{Fit linear mixed nodels for all features of a pimRes object}
\usage{
fitLMMsSingle(
  obj,
  pi,
  fixedVars,
  randomVars,
  verbose,
  returnModels,
  Formula,
  randomNested,
  features,
  addMoransI,
  weightMats,
  moranFormula
)
}
\arguments{
\item{obj}{A results object, for nearest neighbour distances,
the result of a call to \link{addWeightFunction}}

\item{pi}{character string indicating the desired PI as outcome}

\item{fixedVars}{Names of fixed effects}

\item{randomVars}{Names of random variables, possibly in a single vector to
reflect nesting structure, see details.}

\item{verbose}{A boolean, should the formula be printed?}

\item{returnModels}{a boolean: should the full models be returned?
Otherwise only summary statistics are returned}

\item{Formula}{A formula; if not supplied it will be constructed
from the fixed and random variables}

\item{randomNested}{A boolean, indicating if random effects are nested within
point patterns. See details.}

\item{features}{The features for which to fit linear mixed models.
Defaults to all features in the object}

\item{addMoransI}{A boolean, include Moran's I of the cell-wise PIs in the calculation}

\item{weightMats}{A list of weight matrices for Moran's I}

\item{moranFormula}{Formula for Moran's I model fitting}
}
\value{
A list of test results, if requested also the linear models are returned
}
\description{
Fit linear mixed nodels for all features of a pimRes object
}
\seealso{
\link{buildDataFrame},\link{getResults}
}
