% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitLMMs.R
\name{fitLMMsSingle}
\alias{fitLMMsSingle}
\title{Fit linear mixed nodels for all features of a pimRes object}
\usage{
fitLMMsSingle(
  obj,
  pi,
  fixedVars,
  randomVars,
  verbose,
  returnModels,
  Formula,
  randomNested,
  features,
  addMoransI,
  weightMats,
  moranFormula
)
}
\arguments{
\item{obj}{A results object, for nearest neighbour distances,
the result of a call to \link{addWeightFunction}}

\item{pi}{character string indicating the desired PI as outcome
in the linear mixed model}

\item{fixedVars}{Names of fixed effects}

\item{randomVars}{Names of random variables, possibly in a single vector to
reflect nesting structure, see details.}

\item{verbose}{A boolean, should the formula be printed?}

\item{returnModels}{a boolean: should the full models be returned?
Otherwise only summary statistics are returned}

\item{Formula}{A formula; if not supplied it will be constructed
from the fixed and random variables}

\item{randomNested}{A boolean, indicating if random effects are nested within
point patterns. See details.}

\item{features}{The features for which to fit linear mixed models.
Defaults to all features in the object}

\item{addMoransI}{A boolean, include Moran's I in the calculation}

\item{weightMats}{A list of weight matrices for Moran's I}

\item{moranFormula}{Formula for Moran's I model fitting}
}
\value{
A list of test results, if requested also the linear models are returned
}
\description{
Fit linear mixed nodels for all features of a pimRes object
}
\details{
Genes or gene pairs with insufficient observations will be silently
omitted. When randomVars is provided as a vector, independent random
intercepts are fitted for them by default. Providing them separated by '\' or
':' as in the lmer formulas is also allowed to reflect nesting structure.

It is by default assumed that random effects are nested within the point
 patterns. This means for instance that cells with the same name but from
 different point patterns are assigned to different random effects. Set
 'randomNested' to FALSE to override this behaviour.
}
\seealso{
\link{buildDataFrame},\link{getResults}
}
