% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitLMMs.R
\name{fitSingleLmmModel}
\alias{fitSingleLmmModel}
\title{Take an existing frame, add outcome and weight and fit lmer model}
\usage{
fitSingleLmmModel(ff, y, Control, Terms, modMat, MM, Assign, weights = NULL)
}
\arguments{
\item{ff}{The prepared frame}

\item{y}{outcome vector}

\item{Control}{Control parameters}

\item{MM}{A boolean, should a mixed model be tried}

\item{Assign, Terms}{Added to fitted fixed effects model}

\item{weights}{weights vector}
}
\value{
A fitted lmer model
}
\description{
Take an existing frame, add outcome and weight and fit lmer model
}
