% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliaryFunctions.R
\name{getApproxRanks}
\alias{getApproxRanks}
\title{Approximate the ranks given and ecdf}
\usage{
getApproxRanks(ecdf, obs, n = getN(ecdf))
}
\arguments{
\item{ecdf}{An empirical density function}

\item{obs}{A vector of observations}

\item{n}{The total number of observations in the point pattern}
}
\value{
The approximate ranks
}
\description{
Approximate the ranks given and ecdf
}
