% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCrossDist.R
\name{getCrossDist}
\alias{getCrossDist}
\title{Get a cross distance matrix, avoiding zero entries}
\usage{
getCrossDist(p, pSub, null, id = NULL)
}
\arguments{
\item{p}{The total point pattern}

\item{pSub}{A smaller sampled point pattern, representing the null}

\item{null}{A character vector, indicating how the null distribution is
defined. See details.}

\item{id}{The elements of p present in pSub}
}
\value{
A cross distance matrix without zeroes
}
\description{
Get a cross distance matrix, avoiding zero entries
}
