% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDesignVars.R
\name{getDesignVars}
\alias{getDesignVars}
\title{Return all design variables, both at the level of the point pattern and the
level of the event}
\usage{
getDesignVars(x)
}
\arguments{
\item{x}{The results list}

\item{exclude}{variables to exclude}
}
\value{
A vector of likely design variables
}
\description{
Return all design variables, both at the level of the point pattern and the
level of the event
}
\details{
getDesignVars() returns all design variables, getPPPvars returns
design variables related to the different images and getEventVars returs
design variables related to the individual events
}
