% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDesignVars.R
\name{getEventVars}
\alias{getEventVars}
\title{Extract variables from events (the marks)}
\usage{
getEventVars(x, exclude = c("x", "y", "z"))
}
\arguments{
\item{x}{The results list}
}
\value{
A vector of variables
}
\description{
Extract variables from events (the marks)
}
