% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliaryFunctions.R
\name{getFeatures}
\alias{getFeatures}
\title{Extract all unique features}
\usage{
getFeatures(x)
}
\arguments{
\item{x}{A hyperframe or a results list containing a hyperframe}
}
\value{
A vector of features
}
\description{
Extract all unique features
}
