% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFeatures.R
\name{getFeatures}
\alias{getFeatures}
\title{Extract all unique features from an object, or the ones for which PIs were estimated}
\usage{
getFeatures(x)
}
\arguments{
\item{x}{A hyperframe or a results list containing a hyperframe}
}
\value{
A vector of features
}
\description{
Extract all unique features from an object, or the ones for which PIs were estimated
}
\examples{
data(Yang)
hypYang <- buildHyperFrame(Yang,
    coordVars = c("x", "y"),
    imageVars = c("day", "root", "section")
)
head(getFeatures(hypYang))
}
