% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPoissonPi.R
\name{getPoissonPi}
\alias{getPoissonPi}
\title{Calculate the PI under the CSR null hypothesis,
weighing by Poisson variability}
\usage{
getPoissonPi(NP, nAll, approxRanks)
}
\arguments{
\item{NP}{The number of events}

\item{nAll}{The number of events in the calculation of the PI under the null}

\item{approxRanks}{The approximate ranks in the null distribution}
}
\value{
The PI
}
\description{
Calculate the PI under the CSR null hypothesis,
weighing by Poisson variability
}
