% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractResults.R
\name{getResults}
\alias{getResults}
\title{Extract results for a certain parameter}
\usage{
getResults(obj, parameter)
}
\arguments{
\item{obj}{The result object}

\item{parameter}{The desired parameter}
}
\value{
The matrix with result, with p-values in ascending order
}
\description{
Extract results for a certain parameter
}
