% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getResults.R
\name{getResults}
\alias{getResults}
\title{Extract results for a certain parameter}
\usage{
getResults(obj, parameter)
}
\arguments{
\item{obj}{The result object}

\item{parameter}{The desired parameter}
}
\value{
The matrix with result, with p-values in ascending order
}
\description{
Extract results for a certain parameter
}
\examples{
data(Yang)
hypYang = buildHyperFrame(Yang, coordVars = c("x", "y"),
imageVars = c("day", "root", "section"))
#Fit a subset of features to limit computation time
yangPims = estPims(hypYang, pis = "nn", features = attr(hypYang, "features")[seq_len(20)])
#First build the weight function
yangObj <- addWeightFunction(yangPims, designVars = c("day", "root"))
fittedModels = fitLMMs(yangObj, fixedVars = "day", randomVars = "root")
res = getResults(fittedModels, "Intercept")
head(res)
}
