% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getResults.R
\name{getResults}
\alias{getResults}
\title{Extract test results from the linear model}
\usage{
getResults(obj, parameter)
}
\arguments{
\item{obj}{The result object}

\item{parameter}{The desired parameter}
}
\value{
The matrix with results, with p-values in ascending order
\item{Estimate}{The estimated PI}
\item{se}{The corresponding standard error}
\item{pVal}{The p-value}
\item{pAdj}{The Benjamini-Hichberg adjusted p-value}
}
\description{
Extract effect size estimates and p-values for a certain parameter
from a linear model.
}
\examples{
data(Yang)
hypYang <- buildHyperFrame(Yang,
    coordVars = c("x", "y"),
    imageVars = c("day", "root", "section")
)
# Fit a subset of features to limit computation time
yangPims <- estPis(hypYang, pis = "nn")
# First build the weighting function
yangObj <- addWeightFunction(yangPims, designVars = c("day", "root"))
fittedModels <- fitLMMs(yangObj,
    features = getFeatures(yangObj)[seq_len(10)],
    fixedVars = "day", randomVars = "root",
    pi = "nn"
)
res <- getResults(fittedModels, "Intercept")
head(res)
}
