% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux.R
\name{makeDesignVar}
\alias{makeDesignVar}
\title{Make design variable by combining different design variables}
\usage{
makeDesignVar(x, designVars, sep = "_")
}
\arguments{
\item{x}{the design matrix}

\item{designVars}{the design variables to be combined}

\item{sep}{The string to separate the components}
}
\value{
a vector of design levels
}
\description{
Make design variable by combining different design variables
}
