% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux.R
\name{makePairs}
\alias{makePairs}
\title{An aux function to build gene pairs}
\usage{
makePairs(genes)
}
\arguments{
\item{genes}{The genes to be combined}
}
\value{
A chcracter vector of gene pairs
}
\description{
An aux function to build gene pairs
}
