% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliaryFunctions.R
\name{nestRandom}
\alias{nestRandom}
\title{Nest random effects within point patterns, by pasting the design factor
onto it}
\usage{
nestRandom(df, randomVars)
}
\arguments{
\item{df}{The dataframe}

\item{randomVars}{The random variables}
}
\value{
The dataframe with adapted randomVars
}
\description{
Nest random effects within point patterns, by pasting the design factor
onto it
}
