% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nncrossFast.R
\name{nncrossFast}
\alias{nncrossFast}
\title{A faster version of the nncross function in spatstat.geom, by skipping the dmax calculation}
\usage{
nncrossFast(
  X,
  Y,
  dmax,
  iX = NULL,
  iY = NULL,
  what = c("dist", "which"),
  ...,
  k = 1,
  sortby = c("range", "var", "x", "y"),
  is.sorted.X = FALSE,
  is.sorted.Y = FALSE,
  metric = NULL
)
}
\arguments{
\item{X, Y}{Point patterns}

\item{dmax}{The maximium distance, precalculated for speed}

\item{iX, iY, what, ..., k, sortby, is.sorted.X, is.sorted.Y, metric}{See ?spatstat.geom::nncross.ppp}
}
\description{
A faster version of the nncross function in spatstat.geom, by skipping the dmax calculation
}
