% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcNNpi.R
\name{nncrossFast}
\alias{nncrossFast}
\title{Fast version of nncross}
\usage{
nncrossFast(p1, p2)
}
\arguments{
\item{p1, p2}{The point patterns}
}
\value{
A vector of nearest neighbour distances
}
\description{
Fast version of nncross
}
