% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCells.R
\name{plotCells}
\alias{plotCells}
\title{Plot the n cells with highest abundance of a feature}
\usage{
plotCells(obj, feature, nCells = 100)
}
\arguments{
\item{obj}{A hyperframe, or an object containing one}

\item{feature}{The feature to be plotted, a character string}

\item{nCells}{An integer, the number of cells to be plotted}
}
\value{
Plots to the plotting window, returns invisible
}
\description{
After testing for within-cell patterns, it may be useful to
look at the cells with the most events for a certain gene. These are plotted
here, but the spatial location of the cells is lost
}
