% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCells.R
\name{plotCells}
\alias{plotCells}
\title{Plot the n cells with highest abundance of a feature}
\usage{
plotCells(
  obj,
  features = getFeatures(obj)[seq_len(3)],
  nCells = 100,
  Cex = 1.5
)
}
\arguments{
\item{obj}{A hyperframe, or an object containing one}

\item{features}{The features to be plotted, a character vector}

\item{nCells}{An integer, the number of cells to be plotted}

\item{Cex}{The point expansion factor}
}
\value{
Plots to the plotting window, returns invisible
}
\description{
After testing for within-cell patterns, it may be useful to
look at the cells with the most events for certain genes. These are plotted
here, but the spatial location of the cells is lost
}
\examples{
example(addCell, "spatrans")
plotCells(hypFrame2, "gene1")
}
