% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotExplore.R
\name{plotExplore}
\alias{plotExplore}
\title{Plot the hyperframe for exploratory purposes}
\usage{
plotExplore(
  hypFrame,
  features = getFeatures(hypFrame)[seq_len(6)],
  ppps,
  maxPlot = 1e+05,
  Cex = 1,
  plotWindows = !is.null(hypFrame$owins)
)
}
\arguments{
\item{hypFrame}{The hyperframe}

\item{features}{A small number of features to be fitted. Defaults to the first 5}

\item{ppps}{The rownames or indices of the point patterns to be plotted.
Defaults to the first 8.}

\item{maxPlot}{The maximum number of events plotted per point pattern}

\item{Cex}{Point amplification factor}

\item{plotWindows}{A boolean, should windows be plotted too?}
}
\value{
Plots a facet of point patterns to output
}
\description{
A subset of 8 point patterns is plotted. This function is mainly
meant for exploratory purposes and confirming correct data read-in.
}
\examples{
data(Yang)
hypYang <- buildHyperFrame(Yang,
    coordVars = c('x', 'y'),
    imageVars = c('day', 'root', 'section')
)
plotExplore(hypYang)
plotExplore(hypYang, features = c("SmRBRb", "SmTMO5b", "SmWER--SmAHK4f"))
}
