% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTopResults.R
\name{plotTopResults}
\alias{plotTopResults}
\title{Plot the most significant findings for a certain PI}
\usage{
plotTopResults(
  hypFrame,
  results,
  pi,
  effect = "Intercept",
  smallPI = TRUE,
  sigLevel = 0.05,
  numFeats = 2,
  piThreshold = switch(effect, Intercept = 0.5, 0),
  ...
)
}
\arguments{
\item{hypFrame}{The hyperframe with the data}

\item{results}{The results frame}

\item{pi}{A character string, specifying the probabilistic index}

\item{effect}{The name of the effect}

\item{smallPI}{A boolean, should features with PI smaller than piThreshold be shown? See details.}

\item{sigLevel}{The significance level}

\item{numFeats}{The number of features to plot}

\item{piThreshold}{The threshold for PI, a minimum effect size}

\item{...}{passed onto plotting functions plotCells or plotExplore}
}
\value{
A plot from plotCells or plotExplore, throws an error when no features meet the criteria
}
\description{
Extract the most significant features for a certain PI and direction of effect,
and plot them using an appropriate function.
}
\details{
If smallPI is set to TRUE, features with aggregation, colocalization and vicinity to cell boundary or centroid are shown.
If smallPI is set to FALSE, features with regularity, antilocalization and remoteness from cell boundary or centroid are shown.
}
\examples{
example(fitLMMs, "spatrans")
plotTopResults(hypYang, lmmModels, "nn")
plotTopResults(hypYang, lmmModels, "nn", effect = "Intercept")
}
\seealso{
\link{plotCells},\link{plotExplore},\link{fitLMMs}
}
