% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotWf.R
\name{plotWf}
\alias{plotWf}
\title{Plot the weighting function: The observation weigth as a function of the number of observations}
\usage{
plotWf(wf, ...)
}
\arguments{
\item{wf}{The fitted scam object}

\item{...}{Passed onto the plot.scam function for 1D splines}
}
\value{
For univariate PI, returns a line plot; for bivariate PI a ggplot object
}
\description{
Plot the weighting function: The observation weigth as a function of the number of observations
}
\examples{
data(Yang)
hypYang = buildHyperFrame(Yang, coordVars = c("x", "y"),
designVar = c("day", "root", "section"))
yangPims = estPims(hypYang, pis = c("nn", "nnPair"))
#First Build the weight function
wf <- buildWeightFunction(yangPims, pi = "nn", hypFrame = hypYang, designVars = c("day", "root"))
plotWf(wf)
wfPair <- buildWeightFunction(yangPims, pi = "nnPair", hypFrame = hypYang,
 designVars = c("day", "root"))
plotWf(wfPair)
}
