% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotWf.R
\name{plotWf}
\alias{plotWf}
\title{Plot the variance weighting function}
\usage{
plotWf(obj, pi = obj$pis[1])
}
\arguments{
\item{obj}{A results object, for nearest neighbour distances,
the result of a call to \link{addWeightFunction}}

\item{pi}{The PI for which to plot the weighting function}
}
\value{
For univariate PI, returns a line plot; for bivariate PI a ggplot object
}
\description{
The observation weights are plotted as a function of number of events.
For a univariate PI, this is a line plot, for a bivariate PI this is a
scatterplot of majority gene pair as a function of minority gene pair, with the weight represented as a colour scale
}
\examples{
data(Yang)
hypYang <- buildHyperFrame(Yang,
    coordVars = c("x", "y"),
    imageVars = c("day", "root", "section")
)
yangPims <- estPis(hypYang[c(seq_len(5), seq(25, 29)), ],
    nPointsAll = 1e3,
    pis = c("nn", "nnPair")
)
# First Build the weighting function
yangPims <- addWeightFunction(yangPims, designVars = c("day", "root"))
plotWf(yangPims, "nn")
plotWf(yangPims, "nnPair")
}
