% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotWf.R
\name{plotWf}
\alias{plotWf}
\title{Plot the variance weighting function}
\usage{
plotWf(obj, pi = obj$pis[1])
}
\arguments{
\item{obj}{A results object. For distances to fixed objects, the result of a call to \link{estPis}();
for nearest neighbour distances, the result of a call to \link{addWeightFunction}}

\item{pi}{The PI for which to plot the weighting function}
}
\value{
For univariate PI, returns a line plot; for bivariate PI a ggplot object
}
\description{
The observation weights are plotted as a function of number of events.
For a univariate PI, this is a line plot, for a bivariate PI this is a
scatterplot of majority gene as a function of minority gene, with the weight represented as a colour scale.
The minority respectively majority gene are the genes in the gene pair with least and most events
}
\examples{
example(addWeightFunction, "smoppix")
plotWf(yangObj, "nn")
}
