% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotWf.R
\name{plotWf}
\alias{plotWf}
\title{Plot the weighting function: The observation weigth as a function of the number of observations}
\usage{
plotWf(resList, pi = c("nn", "nnPair", "allDist", "allDistPair"), ...)
}
\arguments{
\item{resList}{The result of a call to estPims()}

\item{pi}{The PI for which to plot the weighting function}

\item{...}{Passed onto the plot.scam function for 1D splines}
}
\value{
For univariate PI, returns a line plot; for bivariate PI a ggplot object
}
\description{
Plot the weighting function: The observation weigth as a function of the number of observations
}
\examples{
data(Yang)
hypYang = buildHyperFrame(Yang, coordVars = c("x", "y"),
imageVars = c("day", "root", "section"))
yangPims = estPims(hypYang, pis = c("nn", "nnPair"), features = attr(hypYang, "features")[1:12])
#First Build the weight function
yangObj <- addWeightFunction(yangPims, designVars = c("day", "root"))
plotWf(yangObj, "nn")
plotWf(yangObj, "nnPair")
}
