% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotWf.R
\name{plotWf}
\alias{plotWf}
\title{Plot the weighting function: The observation weigth as a function of the number of observations}
\usage{
plotWf(wf, ...)
}
\arguments{
\item{wf}{The fitted scam object}

\item{...}{Passed onto the plot.scam function for 1D splines}

\item{se}{A boolean, should standard errors be plotted?}
}
\description{
Plot the weighting function: The observation weigth as a function of the number of observations
}
