% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatrans-package.R
\docType{package}
\name{spatrans}
\alias{spatrans-package}
\alias{spatrans}
\title{spatrans: Analyze Spatial Transcriptomics and Other Spatial Omics Data on the Single Molecule Scale}
\description{
Test for univariate and multivariate spatial patterns in spatial omics data with single-molecule resolution. The tests implemented allow for analysis of nested designs and are automatically calibrated to different biological specimens. Tests for aggregation, colocalization and vicinity to cell edge are provided.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/sthawinke/spatrans}
  \item Report bugs at \url{https://github.com/sthawinke/spatrans}
}

}
\author{
\strong{Maintainer}: Stijn Hawinkel \email{stijn.hawinkel@psb.ugent.be} (\href{https://orcid.org/0000-0002-4501-5180}{ORCID})

}
\keyword{internal}
