% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subSampleP.R
\name{subSampleP}
\alias{subSampleP}
\title{Subsample a point pattern when it is too large}
\usage{
subSampleP(p, nSims)
}
\arguments{
\item{p}{The point pattern}

\item{nSims}{The maximum size}
}
\value{
A point pattern, subsample if necessary
}
\description{
Subsample a point pattern when it is too large
}
