% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliaryFunctions.R
\name{sund}
\alias{sund}
\title{Helper function to spit gene pairs}
\usage{
sund(x, sep = "--")
}
\arguments{
\item{x}{character string}

\item{sep}{The character used to split}
}
\value{
The split string
}
\description{
Helper function to spit gene pairs
}
\examples{
GenePair <- 'gene1--gene2'
sund(GenePair)
}
