% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Yang}
\alias{Yang}
\title{Spatial transcriptomics data of Selaginella moellendorffii roots}
\format{
A data matrix
\describe{
  \item{x,y}{Molecule coordinates}
  \item{gene}{Character vector with gene identities}
 \item{root,section,day}{Design variables}
}
}
\source{
\doi{10.1016/j.cub.2023.08.030}
}
\usage{
data(Yang)
}
\description{
Single-molecule spatial transcriptomics smFISH data of Selaginella moellendorffii roots of a
replicated experiment by \insertCite{Yang2023}{spatrans}. Molecule locations, gene identity and design
variables are included. Only a subset of the data, consisting of roots 1-3 and sections 1-5 is included for computational and memory reasons.
}
\references{
\insertAllCited{}
}
\keyword{datasets}
